refactor(linker): split impl Linker into per-concern files
Following the file split in #703, lib.rs still held a 2300-line
`impl Linker` block. Carved it across three files. lib.rs drops to
239 lines — only types, struct definition, module decls, and
re-exports. No behaviour change; methods reach each other across the
split via the same crate-internal visibility they had before.

- builder.rs: new + new_with_gvs + all with_* + simple accessors.
  Adds two tiny helpers (push_glob_patterns, matches_with_negations)
  to dedupe the public-hoist / hoist glob-list construction +
  matching that was previously open-coded twice.
- materialize.rs: detect_strategy(_cross), ensure_in_virtual_store,
  ensure_in_aube_dir, materialize_into, link_file_fresh, plus
  classify_link_error, invalidate_stale_index_for_package,
  validate_index_key.
- link.rs: link_all, link_workspace_hoisted, link_workspace,
  link_hidden_hoist(_at), hoist_remaining_into, plus
  reconcile_top_level_link, build_nested_link_targets.
