+++ crates/aube-linker/src/lib.rs
@@ -1,16 +1,19 @@
-use tracing::{debug, trace, warn};
-
-use aube_lockfile::dep_path_filename::{
-    DEFAULT_VIRTUAL_STORE_DIR_MAX_LENGTH, dep_path_to_filename,
-};
 use aube_lockfile::graph_hash::GraphHashes;
-use aube_lockfile::{LocalSource, LockedPackage, LockfileGraph};
-use aube_store::{PackageIndex, Store, StoredFile};
-use std::collections::BTreeMap;
-use std::path::{Path, PathBuf};
+use aube_store::Store;
+use std::path::PathBuf;
 
+#[cfg(test)]
+use aube_store::PackageIndex;
+#[cfg(test)]
+use std::collections::BTreeMap;
+#[cfg(test)]
+use std::path::Path;
+
+mod builder;
 mod error;
 mod hoisted;
+mod link;
+mod materialize;
 mod patches;
 mod pool;
 mod sweep;
@@ -23,18 +26,13 @@ mod tests;
 
 pub use error::Error;
 pub use hoisted::HoistedPlacements;
+pub use link::build_nested_link_targets;
+pub(crate) use materialize::{invalidate_stale_index_for_package, validate_index_key};
 pub use patches::Patches;
-pub(crate) use patches::{
-    apply_multi_file_patch, current_patch_hashes, read_applied_patches,
-    wipe_changed_patched_entries, write_applied_patches,
-};
+pub(crate) use patches::apply_multi_file_patch;
 pub use pool::default_linker_parallelism;
-pub(crate) use pool::with_link_pool;
-pub(crate) use sweep::{
-    EntryState, classify_entry_state, remove_hidden_hoist_tree, sweep_dead_hidden_hoist_entries,
-    sweep_stale_top_level_entries, try_remove_entry,
-};
 pub use sweep::{is_physical_importer, mkdirp, remove_dir_all_with_retry, sweep_stale_tmp_dirs};
+pub(crate) use sweep::{sweep_stale_top_level_entries, try_remove_entry};
 pub use sys::{
     BinShimOptions, create_bin_shim, create_dir_link, normalize_path, parse_posix_shim_target,
     remove_bin_shim, validate_bin_name, validate_bin_target,
@@ -223,2344 +221,6 @@ pub enum LinkStrategy {
     Copy,
 }
 
-impl Linker {
-    pub fn new(store: &Store, strategy: LinkStrategy) -> Self {
-        Self::new_with_gvs(store, strategy, !aube_util::env::is_ci())
-    }
-
-    /// Select the layout mode. Defaults to `NodeLinker::Isolated`
-    /// (pnpm's `.aube/`-backed virtual-store layout); `Hoisted`
-    /// dispatches `link_all` / `link_workspace` to the flat
-    /// node_modules materializer in `crate::hoisted`.
-    pub fn with_node_linker(mut self, node_linker: NodeLinker) -> Self {
-        self.node_linker = node_linker;
-        self
-    }
-
-    /// Current layout mode. The install driver reads this after
-    /// linking to decide how to resolve per-package directories for
-    /// bin linking and lifecycle scripts — isolated uses the
-    /// `.aube/<dep_path>` convention, hoisted consults the
-    /// `HoistedPlacements` returned on `LinkStats`.
-    pub fn node_linker(&self) -> NodeLinker {
-        self.node_linker
-    }
-
-    /// Override the name of the project-level `node_modules` directory
-    /// (pnpm's `modules-dir` setting). Empty strings are coerced back
-    /// to the default so a `.npmrc` typo can't make the linker write
-    /// into the project root itself. The setting only affects the
-    /// outer directory name — the inner virtual-store layout still
-    /// uses the literal `node_modules` that Node's resolver expects
-    /// when walking up from inside a package.
-    pub fn with_modules_dir_name(mut self, name: impl Into<String>) -> Self {
-        let s = name.into();
-        self.modules_dir_name = if s.trim().is_empty() {
-            "node_modules".to_string()
-        } else {
-            s
-        };
-        self
-    }
-
-    /// Project-level modules directory name. `aube` reads this
-    /// when it needs the same path the linker writes into — keeping
-    /// the computation DRY with whatever the linker was built with.
-    pub fn modules_dir_name(&self) -> &str {
-        &self.modules_dir_name
-    }
-
-    /// Override the per-project virtual-store path (pnpm's
-    /// `virtualStoreDir`). The supplied path should be *absolute* —
-    /// `aube` resolves relative `.npmrc` / `pnpm-workspace.yaml`
-    /// values against the project dir before handing them here.
-    /// When not set, the linker derives the virtual store path as
-    /// `<project_dir>/<modules_dir_name>/.aube` at link time, which
-    /// matches the historical behavior.
-    pub fn with_aube_dir_override(mut self, path: std::path::PathBuf) -> Self {
-        self.aube_dir_override = Some(path);
-        self
-    }
-
-    /// Compute the effective `.aube/` path for `project_dir`.
-    /// Consults the override installed by `with_aube_dir_override` if
-    /// any; otherwise falls back to `<project_dir>/<modules_dir>/.aube`.
-    /// Used internally by `link_all`; also called by the install
-    /// driver's "already linked" fast path so both sites land on the
-    /// same directory when the user has overridden `virtualStoreDir`.
-    pub fn aube_dir_for(&self, project_dir: &Path) -> std::path::PathBuf {
-        self.aube_dir_override
-            .clone()
-            .unwrap_or_else(|| project_dir.join(&self.modules_dir_name).join(".aube"))
-    }
-
-    /// Override the package-level linker worker count. Values below 1
-    /// are ignored by the install driver before they reach this point.
-    pub fn with_link_concurrency(mut self, concurrency: Option<usize>) -> Self {
-        self.link_concurrency = concurrency;
-        self
-    }
-
-    /// Override the global-virtual-store toggle set by `Linker::new`
-    /// (which looks at `CI`). Callers use this to force per-project
-    /// materialization when they've detected a consumer that breaks on
-    /// directory symlinks escaping the project root — e.g. Next.js /
-    /// Turbopack, which canonicalizes `node_modules/<pkg>` and rejects
-    /// anything that lands outside its declared filesystem root.
-    pub fn with_use_global_virtual_store(mut self, enabled: bool) -> Self {
-        self.use_global_virtual_store = enabled;
-        self
-    }
-
-    fn link_parallelism(&self) -> usize {
-        self.link_concurrency
-            .unwrap_or_else(default_linker_parallelism)
-            .max(1)
-    }
-
-    /// Enable pnpm's `shamefully-hoist` mode. When true, every package
-    /// in the graph gets a top-level `node_modules/<name>` symlink in
-    /// addition to the direct-dep entries, producing npm's flat
-    /// layout at the cost of phantom-dep correctness. First-write-wins
-    /// on duplicate names, so root deps always take precedence.
-    pub fn with_shamefully_hoist(mut self, shamefully_hoist: bool) -> Self {
-        self.shamefully_hoist = shamefully_hoist;
-        self
-    }
-
-    /// Configure pnpm's `public-hoist-pattern`. Each input is a glob
-    /// matched against package names; a leading `!` flips it into a
-    /// negation. After the usual direct-dep symlinks, every non-local
-    /// package whose name matches at least one positive pattern and
-    /// no negation gets a top-level `node_modules/<name>` symlink.
-    /// Invalid patterns are silently dropped (same tolerance as
-    /// pnpm), so a typo in `.npmrc` degrades to "not hoisted" instead
-    /// of failing the install.
-    pub fn with_public_hoist_pattern(mut self, patterns: &[String]) -> Self {
-        for raw in patterns {
-            let (neg, body) = match raw.strip_prefix('!') {
-                Some(rest) => (true, rest),
-                None => (false, raw.as_str()),
-            };
-            let Ok(pat) = glob::Pattern::new(body) else {
-                continue;
-            };
-            if neg {
-                self.public_hoist_negations.push(pat);
-            } else {
-                self.public_hoist_patterns.push(pat);
-            }
-        }
-        self
-    }
-
-    /// Toggle pnpm's `hoist` setting. When true (the default), the
-    /// hidden modules tree at `node_modules/.aube/node_modules/` is
-    /// populated via `with_hoist_pattern`. When false, that tree is
-    /// skipped and any existing directory is swept so stale symlinks
-    /// from a previous `hoist=true` run don't keep resolving.
-    pub fn with_hoist(mut self, hoist: bool) -> Self {
-        self.hoist = hoist;
-        self
-    }
-
-    /// Configure pnpm's `hoist-pattern`. Each input is a glob matched
-    /// against package names; a leading `!` flips it into a negation.
-    /// Every non-local package in the graph whose name matches at
-    /// least one positive pattern (and no negation) gets a
-    /// `node_modules/.aube/node_modules/<name>` symlink — the hidden
-    /// fallback dir for Node's parent-directory walk. Invalid
-    /// patterns are silently dropped (pnpm parity). Supplying an
-    /// empty list or only-negation list means "hoist nothing";
-    /// leaving this unconfigured keeps the default `*` match.
-    pub fn with_hoist_pattern(mut self, patterns: &[String]) -> Self {
-        self.hoist_patterns.clear();
-        self.hoist_negations.clear();
-        for raw in patterns {
-            let (neg, body) = match raw.strip_prefix('!') {
-                Some(rest) => (true, rest),
-                None => (false, raw.as_str()),
-            };
-            let Ok(pat) = glob::Pattern::new(body) else {
-                continue;
-            };
-            if neg {
-                self.hoist_negations.push(pat);
-            } else {
-                self.hoist_patterns.push(pat);
-            }
-        }
-        self
-    }
-
-    /// Toggle pnpm's `hoist-workspace-packages`. When false, the
-    /// linker skips creating `node_modules/<ws-pkg>` symlinks for
-    /// workspace packages in every importer, including the root.
-    /// Cross-importer `workspace:` deps already resolve through the
-    /// lockfile, so only direct `require('<ws-pkg>')` from a package
-    /// that doesn't declare it stops working. Default true (pnpm
-    /// parity).
-    pub fn with_hoist_workspace_packages(mut self, on: bool) -> Self {
-        self.hoist_workspace_packages = on;
-        self
-    }
-
-    /// Toggle pnpm's `dedupe-direct-deps`. When true, the linker
-    /// skips creating a per-importer `node_modules/<name>` symlink for
-    /// any direct dep whose root importer already declares the same
-    /// package at the same resolved version — Node's parent-directory
-    /// walk from inside the workspace package still resolves the same
-    /// copy via the root-level symlink, so consumer code is
-    /// unaffected. Default false (pnpm parity). No-op under
-    /// `virtualStoreOnly=true` (no per-importer symlink pass runs)
-    /// and under `NodeLinker::Hoisted` (each importer gets an
-    /// independent flat tree — no shared root to dedupe against).
-    pub fn with_dedupe_direct_deps(mut self, on: bool) -> Self {
-        self.dedupe_direct_deps = on;
-        self
-    }
-
-    /// Whether `pkg_name` should be symlinked into the hidden hoist
-    /// tree. Returns false when `hoist == false` regardless of
-    /// patterns, or when no positive pattern matches. Matching is
-    /// case-insensitive, matching pnpm.
-    fn hoist_matches(&self, pkg_name: &str) -> bool {
-        if !self.hoist {
-            return false;
-        }
-        let opts = glob::MatchOptions {
-            case_sensitive: false,
-            require_literal_separator: false,
-            require_literal_leading_dot: false,
-        };
-        if !self
-            .hoist_patterns
-            .iter()
-            .any(|p| p.matches_with(pkg_name, opts))
-        {
-            return false;
-        }
-        !self
-            .hoist_negations
-            .iter()
-            .any(|p| p.matches_with(pkg_name, opts))
-    }
-
-    /// Whether `pkg_name` should be promoted to the root
-    /// `node_modules` under the configured `public-hoist-pattern`.
-    /// Names with no positive match are rejected; a name that
-    /// matches a positive pattern is still rejected if any negation
-    /// also matches. Matching is case-insensitive.
-    fn public_hoist_matches(&self, pkg_name: &str) -> bool {
-        if self.public_hoist_patterns.is_empty() {
-            return false;
-        }
-        let opts = glob::MatchOptions {
-            case_sensitive: false,
-            require_literal_separator: false,
-            require_literal_leading_dot: false,
-        };
-        if !self
-            .public_hoist_patterns
-            .iter()
-            .any(|p| p.matches_with(pkg_name, opts))
-        {
-            return false;
-        }
-        !self
-            .public_hoist_negations
-            .iter()
-            .any(|p| p.matches_with(pkg_name, opts))
-    }
-
-    /// Override the virtual-store directory name length cap. Primarily
-    /// a hook for tests and for parity with pnpm's
-    /// `virtual-store-dir-max-length` config; most callers should
-    /// leave it at the default.
-    pub fn with_virtual_store_dir_max_length(mut self, max_length: usize) -> Self {
-        self.virtual_store_dir_max_length = max_length;
-        self
-    }
-
-    /// Toggle pnpm's `virtual-store-only`. When enabled, `link_all` /
-    /// `link_workspace` still populate `.aube/<dep_path>/node_modules`
-    /// (and the shared global virtual store under
-    /// `~/.cache/aube/virtual-store/`) but skip the pass that writes
-    /// top-level `node_modules/<name>` symlinks and the hoisting
-    /// passes that target the same directory. No-op under
-    /// `NodeLinker::Hoisted` — that layout is inherently a flat
-    /// top-level materialization.
-    pub fn with_virtual_store_only(mut self, only: bool) -> Self {
-        self.virtual_store_only = only;
-        self
-    }
-
-    /// Whether this linker will skip the top-level `node_modules/<name>`
-    /// symlink pass. Exposed so the install driver can omit root-level
-    /// bin linking and lifecycle-script invocations when the user has
-    /// asked for a virtual-store-only install — both operate on the
-    /// top-level tree that won't exist.
-    pub fn virtual_store_only(&self) -> bool {
-        self.virtual_store_only
-    }
-
-    /// Install a set of pre-computed graph hashes. Every virtual-store
-    /// path the linker constructs after this point will use the
-    /// hashed subdir name for the matching `dep_path`. Callers
-    /// normally derive the hashes once per install via
-    /// `aube_lockfile::graph_hash::compute_graph_hashes` and pass the
-    /// result in here.
-    pub fn with_graph_hashes(mut self, hashes: GraphHashes) -> Self {
-        self.hashes = Some(hashes);
-        self
-    }
-
-    /// Directory name for `dep_path` inside the global virtual store.
-    /// Applies the graph hash (if any) to fold in build state, then
-    /// runs the result through `dep_path_to_filename` so the final
-    /// name is both filesystem-safe and bounded.
-    fn virtual_store_subdir(&self, dep_path: &str) -> String {
-        let hashed = match &self.hashes {
-            Some(h) => h.hashed_dep_path(dep_path),
-            None => dep_path.to_string(),
-        };
-        dep_path_to_filename(&hashed, self.virtual_store_dir_max_length)
-    }
-
-    /// Directory name for `dep_path` inside a project's local
-    /// `node_modules/.aube/`. Same filename-bounding as the global
-    /// store, but without the graph-hash fold — local `.aube/` is
-    /// keyed by dep_path alone because node's resolver walks by
-    /// dep_path and never inspects the shared-store identity.
-    fn aube_dir_entry_name(&self, dep_path: &str) -> String {
-        dep_path_to_filename(dep_path, self.virtual_store_dir_max_length)
-    }
-
-    /// Whether this linker populates the project's `.aube/` entries as
-    /// symlinks into the shared virtual store (true) or materializes a
-    /// per-project copy (false). Callers that want to mutate package
-    /// directories after linking — e.g. running allowBuilds lifecycle
-    /// scripts — need to know because shared-store writes leak across
-    /// projects.
-    pub fn uses_global_virtual_store(&self) -> bool {
-        self.use_global_virtual_store
-    }
-
-    fn new_with_gvs(store: &Store, strategy: LinkStrategy, use_global_virtual_store: bool) -> Self {
-        Self {
-            virtual_store: store.virtual_store_dir(),
-            store: store.clone(),
-            use_global_virtual_store,
-            strategy,
-            patches: Patches::new(),
-            hashes: None,
-            virtual_store_dir_max_length: DEFAULT_VIRTUAL_STORE_DIR_MAX_LENGTH,
-            shamefully_hoist: false,
-            public_hoist_patterns: Vec::new(),
-            public_hoist_negations: Vec::new(),
-            hoist: true,
-            hoist_patterns: vec![glob::Pattern::new("*").expect("'*' is a valid glob pattern")],
-            hoist_negations: Vec::new(),
-            hoist_workspace_packages: true,
-            dedupe_direct_deps: false,
-            node_linker: NodeLinker::Isolated,
-            link_concurrency: None,
-            virtual_store_only: false,
-            modules_dir_name: "node_modules".to_string(),
-            aube_dir_override: None,
-        }
-    }
-
-    /// Install a set of patch contents to apply at materialize time.
-    /// Replaces any previously installed patches. Pair with
-    /// `with_graph_hashes` whose `patch_hash` callback returns the same
-    /// per-`(name, version)` digest, so the patched bytes land at a
-    /// distinct virtual-store path from the unpatched ones.
-    pub fn with_patches(mut self, patches: Patches) -> Self {
-        self.patches = patches;
-        self
-    }
-
-    /// Detect the best linking strategy for the filesystem at the given path.
-    ///
-    /// One-arg form. Probes within one dir. Fine when store and
-    /// project node_modules share the same mount. Use the two-arg
-    /// form for installs where the store lives on a different
-    /// filesystem than the project (USB drives, bind mounts, Docker
-    /// volumes, cross-drive Windows installs). Otherwise the probe
-    /// reports hardlink based on project-FS self-test, then every
-    /// real link call crosses an FS boundary and hits EXDEV. Runtime
-    /// falls back to `fs::copy` per file silently, thousands of
-    /// wasted syscalls, user thinks they got hardlinks.
-    ///
-    /// Returns `Hardlink` when the probe succeeds, `Copy` otherwise.
-    /// Reflink is reachable only through explicit
-    /// `packageImportMethod = clone` / `clone-or-copy`; `auto` resolves
-    /// to `Hardlink` because hardlink benchmarks faster across every
-    /// target reflink supports (APFS clonefile, btrfs/xfs FICLONE).
-    pub fn detect_strategy(path: &Path) -> LinkStrategy {
-        Self::detect_strategy_cross(path, path)
-    }
-
-    /// Two-arg probe. src is the store shard (or any dir on the
-    /// store FS), dst is the project modules dir (or any dir on the
-    /// destination FS). Probe creates a real cross-mount src file
-    /// and tries to hardlink into dst, which catches EXDEV up front.
-    /// Returns `Hardlink` when the probe succeeds, `Copy` otherwise.
-    pub fn detect_strategy_cross(src_dir: &Path, dst_dir: &Path) -> LinkStrategy {
-        // Memoize per (src_dir, dst_dir) for the process lifetime.
-        // The probe writes a real test file and tries hardlink,
-        // ~2 syscalls + 2 unlinks. Multiple Linker instances within
-        // one install (prewarm + final + per-workspace) all repeat
-        // the probe; cache the answer.
-        type ProbeKey = (std::path::PathBuf, std::path::PathBuf);
-        static CACHE: std::sync::OnceLock<
-            std::sync::RwLock<std::collections::HashMap<ProbeKey, LinkStrategy>>,
-        > = std::sync::OnceLock::new();
-        let key = (src_dir.to_path_buf(), dst_dir.to_path_buf());
-        let cache = CACHE.get_or_init(Default::default);
-        if let Some(hit) = cache.read().expect("probe cache poisoned").get(&key) {
-            return *hit;
-        }
-
-        let test_src = src_dir.join(".aube-link-test-src");
-        let test_dst = dst_dir.join(".aube-link-test-dst");
-
-        let strategy = if std::fs::write(&test_src, b"test").is_ok() {
-            let result = if std::fs::hard_link(&test_src, &test_dst).is_ok() {
-                LinkStrategy::Hardlink
-            } else {
-                LinkStrategy::Copy
-            };
-            let _ = std::fs::remove_file(&test_src);
-            let _ = std::fs::remove_file(&test_dst);
-            result
-        } else {
-            LinkStrategy::Copy
-        };
-
-        // First-write-wins via `entry().or_insert`. Two concurrent
-        // linker probes (prewarm + final) sharing the same
-        // (src_dir, dst_dir) can race on the test files: one observes
-        // hardlink-ok, the other sees the first writer's leftover and
-        // falls back to Copy. `.insert()` would let the wrong Copy
-        // result clobber the correct Hardlink for the rest of the
-        // process; `or_insert` keeps whichever value landed first.
-        *cache
-            .write()
-            .expect("probe cache poisoned")
-            .entry(key)
-            .or_insert(strategy)
-    }
-
-    /// Link all packages into node_modules for the given project.
-    pub fn link_all(
-        &self,
-        project_dir: &Path,
-        graph: &LockfileGraph,
-        package_indices: &BTreeMap<String, PackageIndex>,
-    ) -> Result<LinkStats, Error> {
-        if matches!(self.node_linker, NodeLinker::Hoisted) {
-            let mut stats = LinkStats::default();
-            let mut placements = HoistedPlacements::default();
-            hoisted::link_hoisted_importer(
-                self,
-                project_dir,
-                graph.root_deps(),
-                graph,
-                package_indices,
-                &mut stats,
-                &mut placements,
-            )?;
-            // Hoisted mode doesn't use the isolated `.aube/` virtual
-            // store, so a hidden hoist tree under `.aube/node_modules/`
-            // has no consumer. If a previous isolated install left one
-            // behind, sweep it — hoisted's top-level cleanup preserves
-            // dotfiles, so it wouldn't be removed otherwise, and a
-            // stale tree would keep satisfying phantom deps for any
-            // leftover `.aube/<dep_path>/` directories until their
-            // eventual cleanup. Honors `virtualStoreDir`.
-            let _ = crate::remove_dir_all_with_retry(
-                &self.aube_dir_for(project_dir).join("node_modules"),
-            );
-            stats.hoisted_placements = Some(placements);
-            return Ok(stats);
-        }
-
-        let nm = project_dir.join(&self.modules_dir_name);
-        let aube_dir = self.aube_dir_for(project_dir);
-
-        mkdirp(&aube_dir)?;
-
-        // Reclaim space from prior aborted installs. A crash or
-        // Ctrl+C between materialize_into and the atomic rename
-        // leaves `.tmp-<pid>-*` dirs in the virtual store. Sweep
-        // them now so the current install starts clean.
-        sweep_stale_tmp_dirs(&aube_dir);
-
-        // Clean up stale top-level entries not in the current graph.
-        // With shamefully_hoist, every package name in the graph is
-        // also a legitimate top-level entry, so fold those into the
-        // preserve set before sweeping. Scoped packages live under
-        // `node_modules/@scope/<pkg>`, but `read_dir` on `node_modules`
-        // yields the bare `@scope` directory — so we build a second
-        // set of scope prefixes and preserve any entry that matches.
-        let mut root_dep_names: std::collections::HashSet<&str> =
-            graph.root_deps().iter().map(|d| d.name.as_str()).collect();
-        if self.shamefully_hoist {
-            for pkg in graph.packages.values() {
-                root_dep_names.insert(pkg.name.as_str());
-            }
-        } else if !self.public_hoist_patterns.is_empty() {
-            for pkg in graph.packages.values() {
-                if pkg.local_source.is_none() && self.public_hoist_matches(&pkg.name) {
-                    root_dep_names.insert(pkg.name.as_str());
-                }
-            }
-        }
-        // Preserve the virtual-store leaf name when `aube_dir` sits
-        // directly under `nm`. With the default `.aube` the dotfile
-        // check inside the sweep covers it, but a user who sets
-        // `virtualStoreDir=node_modules/vstore` would otherwise see
-        // the sweep delete the freshly-`mkdirp`d virtual store on
-        // every install because `vstore` isn't a dotfile and isn't
-        // in `root_dep_names`.
-        let aube_dir_leaf: Option<std::ffi::OsString> = if aube_dir.parent() == Some(nm.as_path()) {
-            aube_dir.file_name().map(|s| s.to_owned())
-        } else {
-            None
-        };
-        sweep_stale_top_level_entries(&nm, &root_dep_names, aube_dir_leaf.as_deref());
-
-        let mut stats = LinkStats::default();
-
-        // Reconcile previously-applied patches against the current
-        // `self.patches` set. Without graph hashes (CI / no-global-store
-        // mode) the `.aube/<dep_path>` directory name doesn't change
-        // when a patch is added or removed, so the simple "exists?
-        // skip!" check would otherwise leave stale patched bytes in
-        // place after `aube patch-remove` or fail to apply a brand new
-        // patch after `aube patch-commit`. We track the per-`(name,
-        // version)` patch fingerprint in a sidecar file under
-        // `node_modules/` and wipe the matching `.aube/<dep_path>`
-        // entries whenever the fingerprint changes.
-        let prev_applied = read_applied_patches(&nm);
-        let curr_applied = current_patch_hashes(&self.patches);
-        if !self.use_global_virtual_store {
-            wipe_changed_patched_entries(
-                &aube_dir,
-                graph,
-                &prev_applied,
-                &curr_applied,
-                self.virtual_store_dir_max_length,
-            );
-        }
-
-        let nested_link_targets = build_nested_link_targets(project_dir, graph);
-
-        // Step 1: Populate .aube virtual store
-        //
-        // Local packages (file:/link:) never go into the shared global
-        // virtual store — their source is project-specific, so we
-        // materialize them straight into per-project `.aube/` below.
-        // `link:` entries don't need any `.aube/` entry at all; their
-        // top-level symlink points directly at the target.
-        for (dep_path, pkg) in &graph.packages {
-            let Some(ref local) = pkg.local_source else {
-                continue;
-            };
-            if matches!(local, LocalSource::Link(_)) {
-                continue;
-            }
-            let Some(index) = package_indices.get(dep_path) else {
-                continue;
-            };
-            let aube_entry = aube_dir.join(dep_path);
-            if !aube_entry.exists() {
-                self.materialize_into(
-                    &aube_dir,
-                    dep_path,
-                    pkg,
-                    index,
-                    &mut stats,
-                    false,
-                    nested_link_targets.as_ref(),
-                )?;
-            } else {
-                stats.packages_cached += 1;
-            }
-        }
-
-        if self.use_global_virtual_store {
-            use rayon::prelude::*;
-            use rustc_hash::FxHashSet;
-
-            // Pre-create every parent directory (`aube_dir` itself plus
-            // one entry per unique `@scope/`) once so the per-package
-            // par_iter below does not pay 1.4k `create_dir_all` stat
-            // syscalls. The set is tiny (1-5 entries on a typical
-            // graph) so the serial pre-pass is dwarfed by the wins
-            // inside the par_iter that no longer needs the inner
-            // `mkdirp(parent)` call.
-            let mut step1_parents: FxHashSet<PathBuf> = FxHashSet::default();
-            for (dep_path, pkg) in &graph.packages {
-                if pkg.local_source.is_some() {
-                    continue;
-                }
-                let entry = aube_dir.join(self.aube_dir_entry_name(dep_path));
-                if let Some(parent) = entry.parent() {
-                    step1_parents.insert(parent.to_path_buf());
-                }
-            }
-            for parent in &step1_parents {
-                mkdirp(parent)?;
-            }
-
-            let link_parallelism = self.link_parallelism();
-            let step1_timer = std::time::Instant::now();
-            let step1_results: Vec<Result<LinkStats, Error>> =
-                with_link_pool(link_parallelism, || {
-                    graph
-                        .packages
-                        .par_iter()
-                        .filter_map(|(dep_path, pkg)| {
-                            if pkg.local_source.is_some() {
-                                return None;
-                            }
-                            Some((dep_path, pkg))
-                        })
-                        .map(|(dep_path, pkg)| {
-                            let mut local_stats = LinkStats::default();
-                            let local_aube_entry =
-                                aube_dir.join(self.aube_dir_entry_name(dep_path));
-                            let global_entry =
-                                self.virtual_store.join(self.virtual_store_subdir(dep_path));
-
-                            // Single readlink classifies the entry into one of
-                            // three states and drives the whole per-package
-                            // decision tree below. Avoids the double-check
-                            // (`read_link` then `exists`) the previous version
-                            // did and eliminates the unconditional
-                            // `remove_dir`/`remove_file` pair on cold installs,
-                            // which strace showed as ~1.4k ENOENT syscalls per
-                            // install on the medium fixture.
-                            let state = classify_entry_state(&local_aube_entry, &global_entry);
-
-                            if matches!(state, EntryState::Fresh) {
-                                local_stats.packages_cached += 1;
-                                return Ok(local_stats);
-                            }
-
-                            // Symlink is stale or missing — need the package
-                            // index to (re)materialize. The install driver
-                            // omits `package_indices` entries for packages on
-                            // the fast path; load from the store on demand if
-                            // this one slipped through. This keeps the
-                            // fast-path safe against graph-hash changes that
-                            // invalidate the symlink target (patches, engine
-                            // bumps, `allowBuilds` flips).
-                            let owned_index;
-                            let index = match package_indices.get(dep_path) {
-                                Some(idx) => idx,
-                                None => {
-                                    owned_index = self
-                                        .store
-                                        .load_index(
-                                            pkg.registry_name(),
-                                            &pkg.version,
-                                            pkg.integrity.as_deref(),
-                                        )
-                                        .ok_or_else(|| {
-                                            Error::MissingPackageIndex(dep_path.to_string())
-                                        })?;
-                                    &owned_index
-                                }
-                            };
-                            self.ensure_in_virtual_store(
-                                dep_path,
-                                pkg,
-                                index,
-                                &mut local_stats,
-                                nested_link_targets.as_ref(),
-                            )?;
-
-                            // Only pay the `remove_dir`/`remove_file` syscalls
-                            // when we actually have something to remove.
-                            // On Windows, `.aube/<dep_path>` is an NTFS
-                            // junction (created via `sys::create_dir_link`);
-                            // `remove_file` can't unlink those, so try
-                            // `remove_dir` first and fall back to
-                            // `remove_file` for the unix case (where
-                            // `symlink` produces a file-style link).
-                            if matches!(state, EntryState::Stale) {
-                                let _ = std::fs::remove_dir(&local_aube_entry)
-                                    .or_else(|_| std::fs::remove_file(&local_aube_entry));
-                            }
-                            // Parent dirs were pre-created above the
-                            // par_iter; no per-package `mkdirp` here.
-                            sys::create_dir_link(&global_entry, &local_aube_entry)
-                                .map_err(|e| Error::Io(local_aube_entry.clone(), e))?;
-                            Ok(local_stats)
-                        })
-                        .collect()
-                });
-
-            for result in step1_results {
-                let local_stats = result?;
-                stats.packages_linked += local_stats.packages_linked;
-                stats.packages_cached += local_stats.packages_cached;
-                stats.files_linked += local_stats.files_linked;
-            }
-            tracing::debug!("link:step1 (gvs populate) {:.1?}", step1_timer.elapsed());
-        } else {
-            use rayon::prelude::*;
-
-            // `wipe_changed_patched_entries` above already removed any
-            // `.aube/<dep_path>` whose patch fingerprint changed since
-            // the last install, so the existence check below will fall
-            // through to `materialize_into` for those packages and
-            // pick up the current patch state. In per-project mode the
-            // dep paths are already isolated, so we can materialize
-            // them independently on the same rayon pool the gvs path
-            // uses instead of rebuilding the whole tree serially.
-            let link_parallelism = self.link_parallelism();
-            let step1_results: Vec<Result<LinkStats, Error>> =
-                with_link_pool(link_parallelism, || {
-                    graph
-                        .packages
-                        .par_iter()
-                        .filter_map(|(dep_path, pkg)| {
-                            if pkg.local_source.is_some() {
-                                return None;
-                            }
-                            Some((dep_path, pkg))
-                        })
-                        .map(|(dep_path, pkg)| {
-                            let mut local_stats = LinkStats::default();
-                            let aube_entry = aube_dir.join(self.aube_dir_entry_name(dep_path));
-                            if aube_entry.exists() {
-                                // Already in place from a previous run —
-                                // count as cached. `install.rs`
-                                // deliberately omits this dep_path from
-                                // `package_indices` on the fast path, so
-                                // do the existence check first.
-                                local_stats.packages_cached += 1;
-                                return Ok(local_stats);
-                            }
-                            // Entry missing — load the index. Fast path in
-                            // `install.rs` skips `load_index` when
-                            // `aube_entry` already exists; lazy-load here
-                            // for the case where a patch / allowBuilds
-                            // change invalidated the entry since.
-                            let owned_index;
-                            let index = match package_indices.get(dep_path) {
-                                Some(idx) => idx,
-                                None => {
-                                    owned_index = self
-                                        .store
-                                        .load_index(
-                                            pkg.registry_name(),
-                                            &pkg.version,
-                                            pkg.integrity.as_deref(),
-                                        )
-                                        .ok_or_else(|| {
-                                            Error::MissingPackageIndex(dep_path.to_string())
-                                        })?;
-                                    &owned_index
-                                }
-                            };
-                            self.materialize_into(
-                                &aube_dir,
-                                dep_path,
-                                pkg,
-                                index,
-                                &mut local_stats,
-                                false,
-                                nested_link_targets.as_ref(),
-                            )?;
-                            Ok(local_stats)
-                        })
-                        .collect()
-                });
-
-            for result in step1_results {
-                let local_stats = result?;
-                stats.packages_linked += local_stats.packages_linked;
-                stats.packages_cached += local_stats.packages_cached;
-                stats.files_linked += local_stats.files_linked;
-            }
-        }
-
-        // `virtualStoreOnly=true` skips Steps 2 + 3 — the
-        // user-visible top-level `node_modules/<name>` symlinks and
-        // the hoisting passes that target the same directory — but
-        // Step 4 (the hidden `.aube/node_modules/` hoist) still runs
-        // because that tree lives *inside* the virtual store and
-        // packages walking up for undeclared deps need it. Anything
-        // that walks the user-visible root tree (bin linking,
-        // lifecycle scripts, the state sidecar) is the install
-        // driver's responsibility to skip in this mode.
-        if self.virtual_store_only {
-            self.link_hidden_hoist(&aube_dir, graph)?;
-            if let Err(e) = write_applied_patches(&nm, &curr_applied) {
-                tracing::error!(
-                    code = aube_codes::errors::ERR_AUBE_PATCHES_TRACKING_WRITE,
-                    "failed to write .aube-applied-patches.json: {e}. next install may miss stale patched entries"
-                );
-            }
-            return Ok(stats);
-        }
-
-        // Step 2: Create top-level entries as symlinks into .aube.
-        // The .aube/<dep_path>/node_modules/ directory already contains the
-        // package and sibling symlinks to its direct deps (set up by
-        // materialize_into / ensure_in_virtual_store), so a single symlink at
-        // node_modules/<name> gives Node everything it needs to resolve
-        // transitive deps via its normal directory walk.
-        use rayon::prelude::*;
-
-        let root_deps: Vec<_> = graph.root_deps().to_vec();
-        let link_parallelism = self.link_parallelism();
-        let step2_timer = std::time::Instant::now();
-        let results: Vec<Result<bool, Error>> = with_link_pool(link_parallelism, || {
-            root_deps
-                .par_iter()
-                .map(|dep| {
-                    let target_dir = nm.join(&dep.name);
-
-                    // `link:` direct deps point at the on-disk target with
-                    // a plain symlink, bypassing `.aube/` entirely.
-                    if let Some(pkg) = graph.packages.get(&dep.dep_path)
-                        && let Some(LocalSource::Link(rel)) = pkg.local_source.as_ref()
-                    {
-                        let abs_target = project_dir.join(rel);
-                        let link_parent = target_dir.parent().unwrap_or(&nm);
-                        let rel_target =
-                            pathdiff::diff_paths(&abs_target, link_parent).unwrap_or(abs_target);
-                        if reconcile_top_level_link(&target_dir, &rel_target)? {
-                            return Ok(false);
-                        }
-                        if let Some(parent) = target_dir.parent() {
-                            mkdirp(parent)?;
-                        }
-                        sys::create_dir_link(&rel_target, &target_dir)
-                            .map_err(|e| Error::Io(target_dir.clone(), e))?;
-                        return Ok(true);
-                    }
-
-                    // Verify the source actually exists in .aube before symlinking
-                    let source_dir = aube_dir
-                        .join(self.aube_dir_entry_name(&dep.dep_path))
-                        .join("node_modules")
-                        .join(&dep.name);
-                    if !source_dir.exists() {
-                        return Ok(false);
-                    }
-
-                    // Symlink target is relative to node_modules/<name>'s parent.
-                    // For non-scoped packages the parent is node_modules/, but for
-                    // scoped packages (e.g. @scope/name) it is node_modules/@scope/,
-                    // so we must compute the relative path dynamically.
-                    let link_parent = target_dir.parent().unwrap_or(&nm);
-                    let rel_target = pathdiff::diff_paths(&source_dir, link_parent)
-                        .unwrap_or_else(|| source_dir.clone());
-                    // Target-aware reconcile: a version upgrade keeps the
-                    // old `node_modules/<name>` symlink but it now points
-                    // at a stale `.aube/<old-dep-path>`; we need to
-                    // rewrite it to the new `.aube/<new-dep-path>`.
-                    if reconcile_top_level_link(&target_dir, &rel_target)? {
-                        return Ok(false);
-                    }
-                    if let Some(parent) = target_dir.parent() {
-                        mkdirp(parent)?;
-                    }
-
-                    sys::create_dir_link(&rel_target, &target_dir)
-                        .map_err(|e| Error::Io(target_dir.clone(), e))?;
-
-                    trace!("top-level: {}", dep.name);
-                    Ok(true)
-                })
-                .collect()
-        });
-
-        for result in results {
-            if result? {
-                stats.top_level_linked += 1;
-            }
-        }
-        tracing::debug!(
-            "link:step2 (top-level symlinks) {:.1?}",
-            step2_timer.elapsed()
-        );
-
-        // Step 3: public-hoist-pattern matches get surfaced to the
-        // root first, then shamefully_hoist (if enabled) sweeps up
-        // everything else. Both use first-write-wins so direct deps
-        // keep their symlinks and the pattern-matched names take
-        // precedence over the bulk hoist.
-        if !self.public_hoist_patterns.is_empty() {
-            self.hoist_remaining_into(
-                &nm,
-                &aube_dir,
-                graph,
-                &mut stats,
-                "public-hoist",
-                &|name| self.public_hoist_matches(name),
-            )?;
-        }
-        if self.shamefully_hoist {
-            self.hoist_remaining_into(&nm, &aube_dir, graph, &mut stats, "hoist", &|_| true)?;
-        }
-
-        // Step 4: populate (or sweep) the hidden modules tree under
-        // `.aube/node_modules/`. This runs regardless of the root
-        // hoist passes above — it targets a different consumer
-        // (packages inside the virtual store walking up for
-        // undeclared deps) and wouldn't interact with the
-        // root-level symlinks even on name clashes.
-        self.link_hidden_hoist(&aube_dir, graph)?;
-
-        if let Err(e) = write_applied_patches(&nm, &curr_applied) {
-            tracing::error!(
-                code = aube_codes::errors::ERR_AUBE_PATCHES_TRACKING_WRITE,
-                "failed to write .aube-applied-patches.json: {e}. next install may miss stale patched entries"
-            );
-        }
-        Ok(stats)
-    }
-
-    /// Hoisted-mode workspace linker. Runs the per-importer
-    /// hoisted planner once per importer in the graph, accumulating
-    /// stats + placements into a single `LinkStats`. Each importer
-    /// gets its own independent flat tree (no shared root
-    /// virtual-store like the isolated layout), matching npm
-    /// workspaces and what hoisted-mode toolchains expect: a
-    /// self-contained `node_modules/` under every importer.
-    fn link_workspace_hoisted(
-        &self,
-        root_dir: &Path,
-        graph: &LockfileGraph,
-        package_indices: &BTreeMap<String, PackageIndex>,
-        workspace_dirs: &BTreeMap<String, PathBuf>,
-    ) -> Result<LinkStats, Error> {
-        let mut stats = LinkStats::default();
-        let mut placements = HoistedPlacements::default();
-        for (importer_path, deps) in &graph.importers {
-            if !is_physical_importer(importer_path) {
-                continue;
-            }
-            let importer_dir = if importer_path == "." {
-                root_dir.to_path_buf()
-            } else {
-                // Collapse `..` segments lexically — a parent-relative
-                // importer key (`../sibling`, possible when
-                // `pnpm-workspace.yaml#packages` uses `../**`) needs
-                // to land at the actual sibling dir before
-                // `pathdiff`/`strip_prefix` see it.
-                aube_util::path::normalize_lexical(&root_dir.join(importer_path))
-            };
-            // Workspace deps resolve through `workspace_dirs` rather
-            // than going through the placement tree, so the hoisted
-            // planner shouldn't try to copy their contents. Filter
-            // them out of the seed set — we'll symlink them in a
-            // post-pass below.
-            //
-            // Same gating as the isolated mode below: the resolver
-            // omits a `LockedPackage` for workspace-resolved siblings,
-            // so a name match plus a missing package entry is the
-            // signal that the resolver picked the sibling. When the
-            // resolved package IS in `graph.packages`, the resolver
-            // pinned a registry version and the dep should follow the
-            // normal hoisted-placement path (otherwise the post-pass
-            // would silently substitute the local copy).
-            let planner_deps: Vec<aube_lockfile::DirectDep> = deps
-                .iter()
-                .filter(|d| {
-                    !workspace_dirs.contains_key(&d.name)
-                        || graph.packages.contains_key(&d.dep_path)
-                })
-                .cloned()
-                .collect();
-            hoisted::link_hoisted_importer(
-                self,
-                &importer_dir,
-                &planner_deps,
-                graph,
-                package_indices,
-                &mut stats,
-                &mut placements,
-            )?;
-
-            // Drop workspace deps in as symlinks, same as isolated mode.
-            let nm = importer_dir.join(&self.modules_dir_name);
-            if !self.hoist_workspace_packages {
-                continue;
-            }
-            for dep in deps {
-                let Some(ws_dir) = workspace_dirs.get(&dep.name) else {
-                    continue;
-                };
-                // See planner_deps gating above: skip deps the
-                // resolver actually pinned to a registry version.
-                if graph.packages.contains_key(&dep.dep_path) {
-                    continue;
-                }
-                let link_path = nm.join(&dep.name);
-                if let Some(parent) = link_path.parent() {
-                    mkdirp(parent)?;
-                }
-                try_remove_entry(&link_path);
-                let link_parent = link_path.parent().unwrap_or(&nm);
-                let target = pathdiff::diff_paths(ws_dir, link_parent).unwrap_or(ws_dir.clone());
-                sys::create_dir_link(&target, &link_path)
-                    .map_err(|e| Error::Io(link_path.clone(), e))?;
-                stats.top_level_linked += 1;
-            }
-        }
-        // Same rationale as the non-workspace hoisted path: sweep any
-        // `.aube/node_modules/` left behind by a prior isolated
-        // install so hoisted's dotfile-preserving cleanup doesn't
-        // leak a stale hidden tree. Honors `virtualStoreDir`.
-        let _ = crate::remove_dir_all_with_retry(&self.aube_dir_for(root_dir).join("node_modules"));
-        stats.hoisted_placements = Some(placements);
-        Ok(stats)
-    }
-
-    /// Link all packages for a workspace (multiple importers).
-    ///
-    /// Creates the shared `.aube/` virtual store at root, then for each workspace
-    /// package creates `node_modules/` with its direct deps linked from the root `.aube/`.
-    /// Workspace packages that depend on each other get symlinks to the package directory.
-    pub fn link_workspace(
-        &self,
-        root_dir: &Path,
-        graph: &LockfileGraph,
-        package_indices: &BTreeMap<String, PackageIndex>,
-        workspace_dirs: &BTreeMap<String, PathBuf>,
-    ) -> Result<LinkStats, Error> {
-        if matches!(self.node_linker, NodeLinker::Hoisted) {
-            return self.link_workspace_hoisted(root_dir, graph, package_indices, workspace_dirs);
-        }
-
-        let root_nm = root_dir.join(&self.modules_dir_name);
-        let aube_dir = self.aube_dir_for(root_dir);
-
-        mkdirp(&aube_dir)?;
-        mkdirp(&root_nm)?;
-
-        let mut stats = LinkStats::default();
-
-        // Patch reconciliation. Mirrors `link_all`'s logic: wipe
-        // `.aube/<dep_path>` for any package whose patch fingerprint
-        // changed between the previous and current install. Only
-        // applies to per-project (non-gvs) mode because the gvs path
-        // already folds patches into the hashed `.aube/<dep_path>`
-        // name via `with_graph_hashes`.
-        let prev_applied = read_applied_patches(&root_nm);
-        let curr_applied = current_patch_hashes(&self.patches);
-        if !self.use_global_virtual_store {
-            wipe_changed_patched_entries(
-                &aube_dir,
-                graph,
-                &prev_applied,
-                &curr_applied,
-                self.virtual_store_dir_max_length,
-            );
-        }
-
-        let nested_link_targets = build_nested_link_targets(root_dir, graph);
-
-        // Step 1a: Materialize local (`file:` dir/tarball) packages
-        // straight into the shared per-project `.aube/`. They never
-        // participate in the global virtual store since their source
-        // is project-specific. `link:` deps get no `.aube/` entry at
-        // all — step 2 symlinks directly to the target.
-        for (dep_path, pkg) in &graph.packages {
-            let Some(ref local) = pkg.local_source else {
-                continue;
-            };
-            if matches!(local, LocalSource::Link(_)) {
-                continue;
-            }
-            let Some(index) = package_indices.get(dep_path) else {
-                continue;
-            };
-            let aube_entry = aube_dir.join(self.aube_dir_entry_name(dep_path));
-            if aube_entry.exists() {
-                stats.packages_cached += 1;
-                continue;
-            }
-            self.materialize_into(
-                &aube_dir,
-                dep_path,
-                pkg,
-                index,
-                &mut stats,
-                false,
-                nested_link_targets.as_ref(),
-            )?;
-        }
-
-        // Step 1b: Populate shared .aube virtual store at root for
-        // registry packages. Mirrors `link_all`'s parallel +
-        // Fresh/Missing/Stale state machine so warm re-runs are a
-        // `readlink` per package instead of a recreate per package.
-        if self.use_global_virtual_store {
-            use rayon::prelude::*;
-            use rustc_hash::FxHashSet;
-
-            // Pre-create every parent directory (`aube_dir` itself plus
-            // one entry per unique `@scope/`) once so the per-package
-            // par_iter below does not pay 1.4k `create_dir_all` stat
-            // syscalls. The set is tiny (1-5 entries on a typical
-            // graph) so the serial pre-pass is dwarfed by the wins
-            // inside the par_iter that no longer needs the inner
-            // `mkdirp(parent)` call.
-            let mut step1_parents: FxHashSet<PathBuf> = FxHashSet::default();
-            for (dep_path, pkg) in &graph.packages {
-                if pkg.local_source.is_some() {
-                    continue;
-                }
-                let entry = aube_dir.join(self.aube_dir_entry_name(dep_path));
-                if let Some(parent) = entry.parent() {
-                    step1_parents.insert(parent.to_path_buf());
-                }
-            }
-            for parent in &step1_parents {
-                mkdirp(parent)?;
-            }
-
-            let link_parallelism = self.link_parallelism();
-            let step1_timer = std::time::Instant::now();
-            let step1_results: Vec<Result<LinkStats, Error>> =
-                with_link_pool(link_parallelism, || {
-                    graph
-                        .packages
-                        .par_iter()
-                        .filter_map(|(dep_path, pkg)| {
-                            if pkg.local_source.is_some() {
-                                return None;
-                            }
-                            Some((dep_path, pkg))
-                        })
-                        .map(|(dep_path, pkg)| {
-                            let mut local_stats = LinkStats::default();
-                            let local_aube_entry =
-                                aube_dir.join(self.aube_dir_entry_name(dep_path));
-                            let global_entry =
-                                self.virtual_store.join(self.virtual_store_subdir(dep_path));
-
-                            let state = classify_entry_state(&local_aube_entry, &global_entry);
-
-                            if matches!(state, EntryState::Fresh) {
-                                local_stats.packages_cached += 1;
-                                return Ok(local_stats);
-                            }
-
-                            let owned_index;
-                            let index = match package_indices.get(dep_path) {
-                                Some(idx) => idx,
-                                None => {
-                                    owned_index = self
-                                        .store
-                                        .load_index(
-                                            pkg.registry_name(),
-                                            &pkg.version,
-                                            pkg.integrity.as_deref(),
-                                        )
-                                        .ok_or_else(|| {
-                                            Error::MissingPackageIndex(dep_path.to_string())
-                                        })?;
-                                    &owned_index
-                                }
-                            };
-                            self.ensure_in_virtual_store(
-                                dep_path,
-                                pkg,
-                                index,
-                                &mut local_stats,
-                                nested_link_targets.as_ref(),
-                            )?;
-
-                            if matches!(state, EntryState::Stale) {
-                                let _ = std::fs::remove_dir(&local_aube_entry)
-                                    .or_else(|_| std::fs::remove_file(&local_aube_entry));
-                            }
-                            // Parent dirs were pre-created above the
-                            // par_iter; no per-package `mkdirp` here.
-                            sys::create_dir_link(&global_entry, &local_aube_entry)
-                                .map_err(|e| Error::Io(local_aube_entry.clone(), e))?;
-                            Ok(local_stats)
-                        })
-                        .collect()
-                });
-
-            for result in step1_results {
-                let local_stats = result?;
-                stats.packages_linked += local_stats.packages_linked;
-                stats.packages_cached += local_stats.packages_cached;
-                stats.files_linked += local_stats.files_linked;
-            }
-            tracing::debug!(
-                "link_workspace:step1 (gvs populate) {:.1?}",
-                step1_timer.elapsed()
-            );
-        } else {
-            use rayon::prelude::*;
-
-            let link_parallelism = self.link_parallelism();
-            let step1_results: Vec<Result<LinkStats, Error>> =
-                with_link_pool(link_parallelism, || {
-                    graph
-                        .packages
-                        .par_iter()
-                        .filter_map(|(dep_path, pkg)| {
-                            if pkg.local_source.is_some() {
-                                return None;
-                            }
-                            Some((dep_path, pkg))
-                        })
-                        .map(|(dep_path, pkg)| {
-                            let mut local_stats = LinkStats::default();
-                            let aube_entry = aube_dir.join(self.aube_dir_entry_name(dep_path));
-                            if aube_entry.exists() {
-                                local_stats.packages_cached += 1;
-                                return Ok(local_stats);
-                            }
-                            let owned_index;
-                            let index = match package_indices.get(dep_path) {
-                                Some(idx) => idx,
-                                None => {
-                                    owned_index = self
-                                        .store
-                                        .load_index(
-                                            pkg.registry_name(),
-                                            &pkg.version,
-                                            pkg.integrity.as_deref(),
-                                        )
-                                        .ok_or_else(|| {
-                                            Error::MissingPackageIndex(dep_path.to_string())
-                                        })?;
-                                    &owned_index
-                                }
-                            };
-                            self.materialize_into(
-                                &aube_dir,
-                                dep_path,
-                                pkg,
-                                index,
-                                &mut local_stats,
-                                false,
-                                nested_link_targets.as_ref(),
-                            )?;
-                            Ok(local_stats)
-                        })
-                        .collect()
-                });
-
-            for result in step1_results {
-                let local_stats = result?;
-                stats.packages_linked += local_stats.packages_linked;
-                stats.packages_cached += local_stats.packages_cached;
-                stats.files_linked += local_stats.files_linked;
-            }
-        }
-
-        // `virtualStoreOnly=true` skips per-importer node_modules
-        // population and the root-level hoisting passes, but the
-        // hidden `.aube/node_modules/` hoist (Step 4 below) still
-        // runs because it lives *inside* the virtual store. Bin
-        // linking and lifecycle scripts for the top-level importers
-        // are the install driver's responsibility to skip in this
-        // mode.
-        if self.virtual_store_only {
-            // Sweep root_nm of any user-visible entries a prior
-            // (non-virtualStoreOnly) install left behind. With the
-            // default `virtualStoreDir`, `.aube/` lives directly
-            // under `root_nm` and must be preserved. Custom
-            // `virtualStoreDir` overrides put `.aube/` outside the
-            // sweep zone already.
-            let aube_dir_leaf: Option<std::ffi::OsString> =
-                if aube_dir.parent() == Some(root_nm.as_path()) {
-                    aube_dir.file_name().map(|s| s.to_owned())
-                } else {
-                    None
-                };
-            if let Ok(entries) = std::fs::read_dir(&root_nm) {
-                for entry in entries.flatten() {
-                    let name = entry.file_name();
-                    let name_str = name.to_string_lossy();
-                    if name_str.starts_with('.') {
-                        continue;
-                    }
-                    if aube_dir_leaf.as_deref() == Some(name.as_os_str()) {
-                        continue;
-                    }
-                    try_remove_entry(&entry.path());
-                }
-            }
-            self.link_hidden_hoist(&aube_dir, graph)?;
-            if let Err(e) = write_applied_patches(&root_nm, &curr_applied) {
-                tracing::error!(
-                    code = aube_codes::errors::ERR_AUBE_PATCHES_TRACKING_WRITE,
-                    "failed to write .aube-applied-patches.json: {e}. next install may miss stale patched entries"
-                );
-            }
-            return Ok(stats);
-        }
-
-        // Precompute root importer's direct deps keyed by name so the
-        // per-importer loop below can short-circuit on `dedupeDirectDeps`
-        // without walking the root's dep list for every child entry.
-        // Empty when the root has no direct deps (lockfile-only workspaces)
-        // or when `dedupeDirectDeps=false` — skipping the build on the
-        // common path avoids an allocation the per-dep check would
-        // never consult.
-        let root_deps_by_name: std::collections::HashMap<&str, &aube_lockfile::DirectDep> =
-            if self.dedupe_direct_deps {
-                graph
-                    .importers
-                    .get(".")
-                    .map(|deps| deps.iter().map(|d| (d.name.as_str(), d)).collect())
-                    .unwrap_or_default()
-            } else {
-                std::collections::HashMap::new()
-            };
-
-        // Step 2a: Per-importer setup — ensure each importer's
-        // `node_modules/` exists and sweep entries no longer in that
-        // importer's direct deps. Cheap serial work (workspace
-        // importers count is small; the expensive symlink syscalls
-        // run in parallel below). For the root importer we also
-        // expand the preserve set with `shamefullyHoist` /
-        // `publicHoistPattern` matches so the hoist passes that run
-        // after Step 2 don't redo work they'd have preserved.
-        let aube_dir_leaf_root: Option<std::ffi::OsString> =
-            if aube_dir.parent() == Some(root_nm.as_path()) {
-                aube_dir.file_name().map(|s| s.to_owned())
-            } else {
-                None
-            };
-
-        for (importer_path, deps) in &graph.importers {
-            if !is_physical_importer(importer_path) {
-                continue;
-            }
-            let nm = if importer_path == "." {
-                root_nm.clone()
-            } else {
-                // Same lexical-normalization rationale as the hoisted
-                // path above: a `../sibling` importer key has to land
-                // at the actual sibling's `node_modules` rather than
-                // `<root>/../sibling/node_modules`, otherwise
-                // `pathdiff` produces a symlink target with the wrong
-                // depth (one extra `..` per uncollapsed segment).
-                aube_util::path::normalize_lexical(
-                    &root_dir.join(importer_path).join(&self.modules_dir_name),
-                )
-            };
-            if importer_path != "." {
-                mkdirp(&nm)?;
-            }
-
-            let mut preserve: std::collections::HashSet<&str> =
-                deps.iter().map(|d| d.name.as_str()).collect();
-            if importer_path == "." {
-                if self.shamefully_hoist {
-                    for pkg in graph.packages.values() {
-                        preserve.insert(pkg.name.as_str());
-                    }
-                } else if !self.public_hoist_patterns.is_empty() {
-                    for pkg in graph.packages.values() {
-                        if pkg.local_source.is_none() && self.public_hoist_matches(&pkg.name) {
-                            preserve.insert(pkg.name.as_str());
-                        }
-                    }
-                }
-            }
-            let aube_leaf_here = if importer_path == "." {
-                aube_dir_leaf_root.as_deref()
-            } else {
-                None
-            };
-            sweep_stale_top_level_entries(&nm, &preserve, aube_leaf_here);
-        }
-
-        // Step 2b: Create top-level symlinks in parallel.
-        // Flatten (importer, dep) pairs so every symlink syscall
-        // runs through the rayon pool — 3k+ serial
-        // `create_dir_link` calls was the second-biggest slice of
-        // the workspace install phase before this change.
-        use rayon::prelude::*;
-
-        #[derive(Clone)]
-        struct Step2Task<'a> {
-            importer_path: &'a str,
-            nm: PathBuf,
-            dep: &'a aube_lockfile::DirectDep,
-        }
-        let tasks: Vec<Step2Task<'_>> = graph
-            .importers
-            .iter()
-            .filter(|(importer_path, _)| is_physical_importer(importer_path))
-            .flat_map(|(importer_path, deps)| {
-                let nm = if importer_path == "." {
-                    root_nm.clone()
-                } else {
-                    // Same lexical-normalization rationale as
-                    // `link_workspace_hoisted` above: parent-relative
-                    // importer keys must collapse before `pathdiff`
-                    // computes the top-level symlink target.
-                    aube_util::path::normalize_lexical(
-                        &root_dir.join(importer_path).join(&self.modules_dir_name),
-                    )
-                };
-                deps.iter().map(move |dep| Step2Task {
-                    importer_path: importer_path.as_str(),
-                    nm: nm.clone(),
-                    dep,
-                })
-            })
-            .collect();
-
-        let link_parallelism = self.link_parallelism();
-        let step2_timer = std::time::Instant::now();
-        let step2_results: Vec<Result<bool, Error>> = with_link_pool(link_parallelism, || {
-            tasks
-                .par_iter()
-                .map(|task| {
-                    let Step2Task {
-                        importer_path,
-                        nm,
-                        dep,
-                    } = task;
-
-                    // `dedupeDirectDeps`: non-root importer dep
-                    // already covered by the root symlink +
-                    // parent-directory walk.
-                    if self.dedupe_direct_deps
-                        && *importer_path != "."
-                        && let Some(root_dep) = root_deps_by_name.get(dep.name.as_str())
-                        && root_dep.dep_path == dep.dep_path
-                    {
-                        return Ok(false);
-                    }
-
-                    let link_path = nm.join(&dep.name);
-
-                    // Workspace dep (`workspace:` protocol or bare
-                    // semver that satisfies the sibling's version):
-                    // link straight into the sibling package dir.
-                    //
-                    // Gate on the resolver's decision, not just the
-                    // name match. The resolver omits a `LockedPackage`
-                    // entry for workspace-resolved siblings (the
-                    // `workspace_packages` branch in resolve.rs only
-                    // pushes a `DirectDep`, never inserts into
-                    // `resolved`), so a `dep_path` with no package
-                    // entry means "resolver picked the sibling". When
-                    // the package IS in `graph.packages`, the resolver
-                    // pinned a registry version — even if a sibling
-                    // shares the name, the user's spec didn't
-                    // satisfy it (e.g. `is-positive: "2.0.0"` with a
-                    // workspace sibling at `3.0.0`). Falling through
-                    // to the registry branch in that case prevents the
-                    // linker from silently substituting an
-                    // incompatible local copy for the resolved
-                    // version recorded in the lockfile.
-                    if workspace_dirs.contains_key(&dep.name)
-                        && !graph.packages.contains_key(&dep.dep_path)
-                    {
-                        let ws_dir = &workspace_dirs[&dep.name];
-                        if !self.hoist_workspace_packages {
-                            return Ok(false);
-                        }
-                        let link_parent = link_path.parent().unwrap_or(nm);
-                        let rel_target =
-                            pathdiff::diff_paths(ws_dir, link_parent).unwrap_or(ws_dir.clone());
-                        if reconcile_top_level_link(&link_path, &rel_target)? {
-                            return Ok(false);
-                        }
-                        if let Some(parent) = link_path.parent() {
-                            mkdirp(parent)?;
-                        }
-                        sys::create_dir_link(&rel_target, &link_path)
-                            .map_err(|e| Error::Io(link_path.clone(), e))?;
-                        return Ok(true);
-                    }
-
-                    // `link:` dep — absolute path relative to `root_dir`.
-                    if let Some(locked) = graph.packages.get(&dep.dep_path)
-                        && let Some(LocalSource::Link(rel)) = locked.local_source.as_ref()
-                    {
-                        let abs_target = root_dir.join(rel);
-                        let link_parent = link_path.parent().unwrap_or(nm);
-                        let rel_target =
-                            pathdiff::diff_paths(&abs_target, link_parent).unwrap_or(abs_target);
-                        if reconcile_top_level_link(&link_path, &rel_target)? {
-                            return Ok(false);
-                        }
-                        if let Some(parent) = link_path.parent() {
-                            mkdirp(parent)?;
-                        }
-                        sys::create_dir_link(&rel_target, &link_path)
-                            .map_err(|e| Error::Io(link_path.clone(), e))?;
-                        return Ok(true);
-                    }
-
-                    // Regular registry dep — symlink to the root
-                    // `.aube/<dep_path>/node_modules/<name>`.
-                    let source_dir = aube_dir
-                        .join(self.aube_dir_entry_name(&dep.dep_path))
-                        .join("node_modules")
-                        .join(&dep.name);
-                    if !source_dir.exists() {
-                        return Ok(false);
-                    }
-                    let link_parent = link_path.parent().unwrap_or(nm);
-                    let rel_target = pathdiff::diff_paths(&source_dir, link_parent)
-                        .unwrap_or_else(|| source_dir.clone());
-                    if reconcile_top_level_link(&link_path, &rel_target)? {
-                        return Ok(false);
-                    }
-                    if let Some(parent) = link_path.parent() {
-                        mkdirp(parent)?;
-                    }
-                    sys::create_dir_link(&rel_target, &link_path)
-                        .map_err(|e| Error::Io(link_path.clone(), e))?;
-                    trace!("workspace top-level: {} -> {}", dep.name, importer_path);
-                    Ok(true)
-                })
-                .collect()
-        });
-        for result in step2_results {
-            if result? {
-                stats.top_level_linked += 1;
-            }
-        }
-        tracing::debug!(
-            "link_workspace:step2 (top-level symlinks) {:.1?}",
-            step2_timer.elapsed()
-        );
-
-        // Hoisting passes run against the *root* importer only —
-        // pnpm never hoists into nested workspace packages. Run the
-        // selective public-hoist-pattern first so matched names take
-        // precedence, then `shamefully_hoist` sweeps up everything
-        // else.
-        if !self.public_hoist_patterns.is_empty() {
-            self.hoist_remaining_into(
-                &root_nm,
-                &aube_dir,
-                graph,
-                &mut stats,
-                "workspace public-hoist",
-                &|name| self.public_hoist_matches(name),
-            )?;
-        }
-        if self.shamefully_hoist {
-            self.hoist_remaining_into(
-                &root_nm,
-                &aube_dir,
-                graph,
-                &mut stats,
-                "workspace hoist",
-                &|_| true,
-            )?;
-        }
-
-        // Hidden hoist is shared across importers, so a single sweep
-        // here is sufficient for the whole workspace.
-        self.link_hidden_hoist(&aube_dir, graph)?;
-
-        if let Err(e) = write_applied_patches(&root_nm, &curr_applied) {
-            tracing::error!(
-                code = aube_codes::errors::ERR_AUBE_PATCHES_TRACKING_WRITE,
-                "failed to write .aube-applied-patches.json: {e}. next install may miss stale patched entries"
-            );
-        }
-        Ok(stats)
-    }
-
-    /// Populate (or sweep) the hidden modules directories at
-    /// `aube_dir/node_modules/<name>` and, in global-virtual-store mode,
-    /// `virtual_store/node_modules/<name>`. When `self.hoist` is
-    /// enabled, walks every non-local package in the graph and creates
-    /// a symlink for names that match `hoist_patterns` into each
-    /// corresponding virtual-store package entry.
-    /// When disabled, wipes the directory so previously-hoisted
-    /// symlinks don't keep resolving through Node's parent walk.
-    ///
-    /// Unlike `hoist_remaining_into`, this writes into a private
-    /// sibling of `.aube/<dep_path>/` rather than the visible root
-    /// `node_modules/`. Packages inside the virtual store (e.g.
-    /// `.aube/react@18/node_modules/react/`) walk up through
-    /// `.aube/node_modules/` during require resolution, which is the
-    /// only consumer of these links — nothing inside the user's own
-    /// `node_modules/<name>` view is affected. In GVS mode, many
-    /// toolchains canonicalize the package path into
-    /// `~/.cache/aube/virtual-store/<hash>/node_modules/<name>`, so we
-    /// mirror the hidden hoist under the shared virtual-store root too.
-    fn link_hidden_hoist(&self, aube_dir: &Path, graph: &LockfileGraph) -> Result<(), Error> {
-        self.link_hidden_hoist_at(aube_dir, aube_dir, graph, false, true)?;
-        if self.use_global_virtual_store {
-            self.link_hidden_hoist_at(
-                &self.virtual_store,
-                &self.virtual_store,
-                graph,
-                true,
-                false,
-            )?;
-        }
-        Ok(())
-    }
-
-    fn link_hidden_hoist_at(
-        &self,
-        hidden_root: &Path,
-        source_root: &Path,
-        graph: &LockfileGraph,
-        use_hashed_subdirs: bool,
-        sweep_stale_entries: bool,
-    ) -> Result<(), Error> {
-        let hidden = hidden_root.join("node_modules");
-        // FxHashSet over the borrowed name (lives for the lockfile graph
-        // lifetime) drops the SipHash overhead and the per-insert
-        // `String` clone the `HashSet<String>` version forced.
-        let mut claimed: rustc_hash::FxHashSet<&str> = rustc_hash::FxHashSet::default();
-        let packages: Vec<_> = if self.hoist {
-            graph
-                .packages
-                .iter()
-                .filter_map(|(dep_path, pkg)| {
-                    if pkg.local_source.is_some() || !self.hoist_matches(&pkg.name) {
-                        return None;
-                    }
-                    // First-writer-wins on name clashes across versions.
-                    // BTree iteration over `graph.packages` gives a
-                    // deterministic tiebreaker across runs.
-                    claimed.insert(pkg.name.as_str()).then_some((dep_path, pkg))
-                })
-                .collect()
-        } else {
-            Vec::new()
-        };
-
-        if !self.hoist {
-            // Previous install may have populated this tree with
-            // hoist=true. Drop entries so Node doesn't keep resolving
-            // phantom deps through the stale symlinks. Project-local
-            // hidden hoist owns the whole tree and can remove it in
-            // one shot; the shared GVS mirror only reclaims broken
-            // entries because live links may belong to another project.
-            if sweep_stale_entries {
-                remove_hidden_hoist_tree(&hidden);
-            } else {
-                sweep_dead_hidden_hoist_entries(&hidden);
-            }
-            return Ok(());
-        }
-        // Wipe before repopulating so a dependency removed from the
-        // graph (or a pattern that no longer matches) doesn't linger.
-        // The shared GVS hidden hoist only prunes broken entries:
-        // removing live cross-project links would make the directory
-        // last-writer-wins for sequential installs.
-        if sweep_stale_entries {
-            remove_hidden_hoist_tree(&hidden);
-        } else {
-            sweep_dead_hidden_hoist_entries(&hidden);
-        }
-        for (dep_path, pkg) in packages {
-            let source_subdir = if use_hashed_subdirs {
-                self.virtual_store_subdir(dep_path)
-            } else {
-                self.aube_dir_entry_name(dep_path)
-            };
-            let source_dir = source_root
-                .join(source_subdir)
-                .join("node_modules")
-                .join(&pkg.name);
-            if !source_dir.exists() {
-                continue;
-            }
-            let target_dir = hidden.join(&pkg.name);
-            if let Some(parent) = target_dir.parent() {
-                mkdirp(parent)?;
-            }
-            let link_parent = target_dir.parent().unwrap_or(&hidden);
-            let rel_target = pathdiff::diff_paths(&source_dir, link_parent)
-                .unwrap_or_else(|| source_dir.clone());
-            if reconcile_top_level_link(&target_dir, &rel_target)? {
-                continue;
-            }
-            sys::create_dir_link(&rel_target, &target_dir)
-                .map_err(|e| Error::Io(target_dir.clone(), e))?;
-            trace!("hidden-hoist: {}", pkg.name);
-            // Intentionally not counted in `stats.top_level_linked`.
-            // That counter reflects the user-visible root
-            // `node_modules/<name>` entries; hidden-hoist symlinks
-            // live under `.aube/node_modules/` and are only reached
-            // via Node's parent-directory walk from inside the
-            // virtual store, not from the user's own code.
-        }
-        Ok(())
-    }
-
-    /// Shared `shamefully_hoist` implementation. For every non-local
-    /// package in the graph, create a symlink at `nm/<pkg.name>`
-    /// pointing at the matching `.aube/<dep_path>/node_modules/<pkg.name>`
-    /// entry.
-    ///
-    /// Two separate "first-write-wins" protections apply:
-    ///
-    /// - **Direct deps always win over hoisted transitives.** Names
-    ///   that appear in `graph.root_deps()` were placed (or
-    ///   deliberately skipped) by Step 2 and must never be overwritten
-    ///   by a hoist pass — that would silently swap `node_modules/foo`
-    ///   from the version the user pinned to whatever transitive
-    ///   happened to sort first.
-    /// - **Within the hoist pass, BTree iteration order is the
-    ///   tiebreaker across versions.** The `claimed` set records
-    ///   names we already hoisted this call so a later iteration with
-    ///   the same name (different `dep_path`) doesn't clobber the
-    ///   first winner.
-    ///
-    /// For everything else the caller gets a *target-aware* reconcile:
-    /// an existing symlink at `nm/<name>` that points at the version
-    /// this iteration wants is kept; one pointing at a stale
-    /// `.aube/<old-dep-path>/` (leftover from a prior install whose
-    /// hoisted version has since changed) is replaced. The old
-    /// plain-`exists?` check here kept stale entries because the
-    /// surrounding linker used to wipe `nm` unconditionally — now that
-    /// we sweep surgically, hoist has to cope with partial priors.
-    ///
-    /// `trace_label` distinguishes the `link_all` vs `link_workspace`
-    /// callers in `-v` output.
-    fn hoist_remaining_into(
-        &self,
-        nm: &Path,
-        aube_dir: &Path,
-        graph: &LockfileGraph,
-        stats: &mut LinkStats,
-        trace_label: &str,
-        select: &dyn Fn(&str) -> bool,
-    ) -> Result<(), Error> {
-        // Root direct-dep names. Populated from the importer map
-        // rather than an opaque "touched by Step 2" signal so a direct
-        // dep that *failed* to place (missing `source_dir.exists()`,
-        // workspace toggle, etc.) still reserves its slot — pnpm
-        // doesn't hoist over a direct dep even when the direct dep
-        // couldn't be installed.
-        let direct_dep_names: std::collections::HashSet<&str> =
-            graph.root_deps().iter().map(|d| d.name.as_str()).collect();
-
-        // FxHashSet over the borrowed name (lives for the lockfile graph
-        // lifetime) drops the SipHash overhead and the per-insert
-        // `String` clone the `HashSet<String>` version forced.
-        let mut claimed: rustc_hash::FxHashSet<&str> = rustc_hash::FxHashSet::default();
-
-        for (dep_path, pkg) in &graph.packages {
-            if pkg.local_source.is_some() {
-                continue;
-            }
-            if !select(&pkg.name) {
-                continue;
-            }
-            // Direct deps always win over hoisting.
-            if direct_dep_names.contains(pkg.name.as_str()) {
-                continue;
-            }
-            // First-writer-wins within the hoist pass: if an earlier
-            // iteration already hoisted this name, later iterations
-            // with the same name don't overwrite it.
-            if !claimed.insert(pkg.name.as_str()) {
-                continue;
-            }
-            let source_dir = aube_dir
-                .join(self.aube_dir_entry_name(dep_path))
-                .join("node_modules")
-                .join(&pkg.name);
-            if !source_dir.exists() {
-                // Don't remove `name` from `claimed` — another
-                // iteration for the same name would also find its
-                // `source_dir` missing (the `.aube` populate phase
-                // runs before hoist for every package), and leaving
-                // the name claimed preserves the existing symlink
-                // (whatever it points at) instead of repeatedly
-                // probing for a materialization that isn't coming.
-                continue;
-            }
-            let target_dir = nm.join(&pkg.name);
-            let link_parent = target_dir.parent().unwrap_or(nm);
-            let rel_target = pathdiff::diff_paths(&source_dir, link_parent)
-                .unwrap_or_else(|| source_dir.clone());
-            if reconcile_top_level_link(&target_dir, &rel_target)? {
-                continue;
-            }
-            if let Some(parent) = target_dir.parent() {
-                mkdirp(parent)?;
-            }
-            sys::create_dir_link(&rel_target, &target_dir)
-                .map_err(|e| Error::Io(target_dir.clone(), e))?;
-            trace!("{trace_label}: {}", pkg.name);
-            stats.top_level_linked += 1;
-        }
-        Ok(())
-    }
-
-    /// Materialize a package in the global virtual store if not already present.
-    ///
-    /// Materialize `dep_path` into the shared global virtual store.
-    ///
-    /// Uses atomic rename to avoid TOCTOU races: materializes into a
-    /// PID-stamped temp directory, then renames into place. If another
-    /// process wins the race, its result is kept and the temp dir is
-    /// cleaned up.
-    ///
-    /// Exposed so the install driver can pipeline GVS population into
-    /// the fetch phase: as each tarball finishes importing into the
-    /// CAS, the driver calls this to reflink the package into its
-    /// `~/.cache/aube/virtual-store/<subdir>` entry. Link step 1 then
-    /// hits the `pkg_nm_dir.exists()` fast path and only creates the
-    /// per-project `.aube/<dep_path>` symlink.
-    pub fn ensure_in_virtual_store(
-        &self,
-        dep_path: &str,
-        pkg: &LockedPackage,
-        index: &PackageIndex,
-        stats: &mut LinkStats,
-        // `link:` transitives the resolver pinned (e.g. via root
-        // `pnpm.overrides`) need their on-disk target so the parent's
-        // sibling symlink doesn't dangle into a non-existent
-        // `.aube/<name>@link+...`. `None` means "no nested links in
-        // this graph" and the materialize hot path stays unchanged.
-        nested_link_targets: Option<&BTreeMap<String, PathBuf>>,
-    ) -> Result<(), Error> {
-        let _diag =
-            aube_util::diag::Span::new(aube_util::diag::Category::Linker, "ensure_in_vstore")
-                .with_meta_fn(|| {
-                    format!(
-                        r#"{{"name":{},"files":{}}}"#,
-                        aube_util::diag::jstr(&pkg.name),
-                        index.len()
-                    )
-                });
-        // Global-store paths always run through the vstore_key map —
-        // when hashes are installed this folds dep-graph + engine
-        // state into the leaf name, so concurrent builds of the same
-        // package against different toolchains don't collide.
-        let subdir = self.virtual_store_subdir(dep_path);
-        let pkg_nm_dir = self
-            .virtual_store
-            .join(&subdir)
-            .join("node_modules")
-            .join(&pkg.name);
-
-        if pkg_nm_dir.exists() {
-            trace!("virtual store hit: {dep_path}");
-            stats.packages_cached += 1;
-            return Ok(());
-        }
-
-        // Materialize into a temp directory, then atomically rename into place
-        // to avoid TOCTOU races between concurrent `aube install` processes.
-        // `subdir` already comes from `dep_path_to_filename`, which
-        // flattens `/` to `+` as part of its escape pass, so it's
-        // already safe to splice into a single path component.
-        let tmp_name = format!(".tmp-{}-{subdir}", std::process::id());
-        let tmp_base = self.virtual_store.join(&tmp_name);
-
-        let result = self.materialize_into(
-            &tmp_base,
-            dep_path,
-            pkg,
-            index,
-            stats,
-            true,
-            nested_link_targets,
-        );
-
-        if result.is_err() {
-            let _ = std::fs::remove_dir_all(&tmp_base);
-            return result;
-        }
-
-        // Atomically move the dep_path entry from the temp dir to the final location.
-        let tmp_entry = tmp_base.join(&subdir);
-        let final_entry = self.virtual_store.join(&subdir);
-
-        // Ensure the parent of the final entry exists (e.g. for scoped packages).
-        if let Some(parent) = final_entry.parent() {
-            mkdirp(parent)?;
-        }
-
-        match aube_util::fs_atomic::rename_with_retry(&tmp_entry, &final_entry) {
-            Ok(()) => {
-                trace!("atomically placed {subdir} in virtual store");
-            }
-            Err(e) if final_entry.exists() => {
-                // Another process won the race — that's fine, use theirs.
-                trace!("lost rename race for {dep_path}, using existing: {e}");
-                // Undo the stats from our materialization since we're discarding it
-                stats.packages_linked = stats.packages_linked.saturating_sub(1);
-                stats.files_linked = stats.files_linked.saturating_sub(index.len());
-                stats.packages_cached += 1;
-                // Lost-race path: our `subdir` is still inside
-                // `tmp_base`, so a full recursive delete is needed.
-                let _ = std::fs::remove_dir_all(&tmp_base);
-                return Ok(());
-            }
-            Err(e) => {
-                let _ = std::fs::remove_dir_all(&tmp_base);
-                return Err(Error::Io(final_entry, e));
-            }
-        }
-
-        // Successful rename: `tmp_base` is now an empty wrapper directory
-        // (its single child was the subdir we just renamed out). Use
-        // `remove_dir` instead of `remove_dir_all` — the latter still
-        // does the full `opendir`/`fdopendir`(fcntl)/`readdir`/`close`
-        // walk even on an empty dir, which dtrace shows as ~6 extra
-        // syscalls per package. At 227 packages that's ~1.4k wasted
-        // syscalls on every cold install.
-        //
-        // `remove_dir` fails with `ENOTEMPTY` if a future change to
-        // `materialize_into` starts dropping extra files into
-        // `tmp_base`. Log at debug so the leak is observable without
-        // being fatal; the worst-case outcome is a stray tmp dir, and
-        // concurrent-writer races already use the full
-        // `remove_dir_all` branch above.
-        if let Err(e) = std::fs::remove_dir(&tmp_base) {
-            debug!(
-                "remove_dir({}) failed, leaving tmp in place: {e}",
-                tmp_base.display()
-            );
-        }
-
-        Ok(())
-    }
-
-    /// Materialize a single package directly into the per-project
-    /// virtual store at `aube_dir/<dep_path>/node_modules/<name>/`.
-    ///
-    /// Idempotent: if the entry already exists, counts as cached and
-    /// returns. Used by the install-time materializer to pipeline the
-    /// link work into the fetch phase under non-GVS mode, so the
-    /// dedicated link phase only has to create top-level
-    /// `node_modules/<name>` symlinks.
-    pub fn ensure_in_aube_dir(
-        &self,
-        aube_dir: &Path,
-        dep_path: &str,
-        pkg: &LockedPackage,
-        index: &PackageIndex,
-        stats: &mut LinkStats,
-        nested_link_targets: Option<&BTreeMap<String, PathBuf>>,
-    ) -> Result<(), Error> {
-        // `materialize_into` batches `create_dir_all` for every parent
-        // it needs, so callers don't have to mkdirp the entry's parent
-        // (which is just `aube_dir` itself, already created by the
-        // materializer driver).
-        let entry = aube_dir.join(self.aube_dir_entry_name(dep_path));
-        if entry.exists() {
-            stats.packages_cached += 1;
-            return Ok(());
-        }
-        self.materialize_into(
-            aube_dir,
-            dep_path,
-            pkg,
-            index,
-            stats,
-            false,
-            nested_link_targets,
-        )
-    }
-
-    /// Materialize a package's files and transitive dep symlinks into a base directory.
-    ///
-    /// `apply_hashes` controls whether per-dep subdir names are run
-    /// through `vstore_key` (the content-addressed name) or used as
-    /// raw `dep_path` strings. Global-store callers pass `true` so
-    /// the shared `~/.cache/aube/virtual-store/` can hold isolated
-    /// copies for each `(deps_hash, engine)` combination;
-    /// per-project `.aube/` callers pass `false` because node's
-    /// runtime module walk resolves by dep_path only.
-    #[allow(clippy::too_many_arguments)]
-    fn materialize_into(
-        &self,
-        base_dir: &Path,
-        dep_path: &str,
-        pkg: &LockedPackage,
-        index: &PackageIndex,
-        stats: &mut LinkStats,
-        apply_hashes: bool,
-        // dep_path → absolute on-disk target for any `link:` packages
-        // referenced as transitive deps. When the parent itself is a
-        // `file:` Directory or `link:` Link (workspace-style locals),
-        // its `package.json` may declare `link:./libs/foo` deps that
-        // point inside the parent's source tree. We sidestep the
-        // virtual store for those — there is no `.aube/<dep>@link+...`
-        // entry — and symlink straight to the on-disk path the
-        // resolver pinned. `None` means "no nested link transitives in
-        // this graph", which is the common case.
-        nested_link_targets: Option<&BTreeMap<String, PathBuf>>,
-    ) -> Result<(), Error> {
-        let subdir = if apply_hashes {
-            self.virtual_store_subdir(dep_path)
-        } else {
-            self.aube_dir_entry_name(dep_path)
-        };
-        let pkg_nm_dir = base_dir.join(&subdir).join("node_modules").join(&pkg.name);
-
-        // Pre-compute the set of unique parent directories across
-        // every file in the index AND every scoped transitive-dep
-        // symlink we're about to create, then mkdir them in a single
-        // pass. Previously each file looped through `mkdirp(parent)`
-        // which always did an `exists()` check (= statx syscall) even
-        // though the same parents were shared by dozens of siblings —
-        // `materialize_into` for a typical 32-file npm package
-        // resulted in ~25 redundant statx calls. Collecting the unique
-        // parents first, sorting by length (so ancestors precede
-        // descendants), and calling `create_dir_all` once each cuts
-        // out the redundant stats entirely. `BTreeSet` sorts
-        // lexicographically, which is good enough because every
-        // ancestor of a directory is a prefix of it.
-        let pkg_nm_parent = base_dir.join(&subdir).join("node_modules");
-        // Collect into Vec + sort + dedup instead of BTreeSet. For a
-        // package with thousands of files (typescript, next), the
-        // BTreeSet's per-insert log-N PathBuf comparison (~50-byte
-        // memcmps) was a measurable cost on top of the redundant
-        // create_dir_all that the set was deduplicating in the first
-        // place.
-        let mut parents: Vec<PathBuf> = Vec::with_capacity(index.len() / 4 + 4);
-        parents.push(pkg_nm_dir.clone());
-        // Validate every key once here. The file-linking loop below
-        // walks the same immutable index, so skipping the check
-        // there is safe.
-        for rel_path in index.keys() {
-            validate_index_key(rel_path)?;
-            let target = pkg_nm_dir.join(rel_path);
-            if let Some(parent) = target.parent() {
-                parents.push(parent.to_path_buf());
-            }
-        }
-        // Scoped transitive deps need `pkg_nm_parent/@scope/` to exist
-        // before the symlink call; include those parents in the batch.
-        for dep_name in pkg.dependencies.keys() {
-            if let Some(slash) = dep_name.find('/')
-                && dep_name.starts_with('@')
-            {
-                parents.push(pkg_nm_parent.join(&dep_name[..slash]));
-            }
-        }
-        parents.sort_unstable();
-        parents.dedup();
-        for parent in &parents {
-            std::fs::create_dir_all(parent).map_err(|e| Error::Io(parent.clone(), e))?;
-        }
-
-        // `materialize_into` always writes into a fresh location
-        // (either a `.tmp-<pid>-...` staging dir for the global virtual
-        // store or a per-project `.aube/<dep_path>` just created by
-        // the caller), so we can skip the `remove_file(dst)` that
-        // `link_file` does defensively. Pass `fresh = true` to suppress
-        // the unlink syscall on every file. For a 1.4k-package install
-        // that's ~45k wasted `unlink` calls on the hot path.
-        for (rel_path, stored) in index {
-            // Key already validated in the parent-collection loop
-            // above. The index is immutable between the two loops.
-            let target = pkg_nm_dir.join(rel_path);
-
-            if let Err(e) = self.link_file_fresh(stored, rel_path, &target) {
-                if let Error::MissingStoreFile { .. } = &e {
-                    invalidate_stale_index_for_package(&self.store, pkg);
-                }
-                return Err(e);
-            }
-            stats.files_linked += 1;
-
-            if stored.executable {
-                // `create_cas_file` writes every CAS entry as 0o644
-                // unconditionally; the only place a CAS entry's
-                // shared inode gets the +x bit is the very first
-                // `make_executable` call against a hardlinked or
-                // reflinked target — that `chmod` upgrades the
-                // shared inode for every later linker that points
-                // at it. Skipping the call (an earlier optimization)
-                // produced 0o644 binaries on cold installs and
-                // broke every CLI shipped via npm.
-                #[cfg(unix)]
-                xx::file::make_executable(&target).map_err(|e| Error::Xx(e.to_string()))?;
-            }
-        }
-
-        // Apply any user-supplied patch for this `(name, version)`.
-        // Patches are applied *after* the files have been linked into
-        // the virtual store but *before* transitive symlinks, so the
-        // patched bytes live alongside the unpatched ones at a
-        // distinct subdir (the graph hash callback is responsible for
-        // making sure that's true).
-        let patch_key = pkg.spec_key();
-        if let Some(patch_text) = self.patches.get(&patch_key) {
-            apply_multi_file_patch(&pkg_nm_dir, patch_text)
-                .map_err(|msg| Error::Patch(patch_key.clone(), msg))?;
-        }
-
-        // Create symlinks for transitive dependencies. Parents for
-        // scoped packages were added to the `parents` batch above, so
-        // we no longer need a per-symlink mkdirp. We also skip the
-        // `symlink_metadata().is_ok()` existence check: callers
-        // guarantee the target directory is freshly created (either a
-        // `.tmp-<pid>-...` staging dir for the global virtual store or
-        // a per-project `.aube/<dep_path>` that the caller just
-        // ensured is empty), so nothing can be in the way.
-        for (dep_name, dep_version) in &pkg.dependencies {
-            let dep_dep_path = format!("{dep_name}@{dep_version}");
-            // Skip any dep whose name matches the package being
-            // materialized, regardless of version. The symlink would
-            // land at `pkg_nm_parent.join(dep_name)` which is exactly
-            // `pkg_nm_dir` — the directory we just populated with the
-            // package's own files — and `create_dir_link` would fail
-            // EEXIST. The skip used to require version-equality too,
-            // but published packages occasionally declare a *different*
-            // version of themselves as a dep (e.g. `react_ujs@3.3.0`
-            // pins `react_ujs@^2.7.1`, an artifact of how its build
-            // script generates its package.json). Treat that as a
-            // self-reference: `require('<self>')` from inside the
-            // package resolves to its own files, matching what npm /
-            // pnpm / yarn end up with after their hoisting passes.
-            if dep_name == &pkg.name {
-                continue;
-            }
-            let symlink_path = pkg_nm_parent.join(dep_name);
-            // `link:` transitive: the resolver pinned an absolute
-            // on-disk target. Skip the virtual-store sibling lookup
-            // (there is no `.aube/<dep>@link+...` entry for these) and
-            // symlink straight at the source directory.
-            //
-            // Store the absolute target verbatim. A relative path
-            // would have to thread two pitfalls at once: the GVS
-            // tmp→final rename (link's own depth changes by one) AND
-            // macOS `/tmp`→`/private/tmp` symlink expansion (the dir
-            // the OS resolves the link from is one level deeper than
-            // `self.virtual_store` lexically suggests). Either alone
-            // is fixable; together every `pathdiff` variant lands one
-            // component off and the link dangles. Sibling symlinks
-            // get away with relative paths because both endpoints
-            // live inside `base_dir` and move together; nested-link
-            // targets are *external* (under `project_dir`) so the
-            // tricks that work for siblings don't apply. Windows
-            // already uses absolute targets for the same reason (see
-            // the `#[cfg(windows)]` block below).
-            if let Some(map) = nested_link_targets
-                && let Some(abs_target) = map.get(&dep_dep_path)
-            {
-                sys::create_dir_link(abs_target, &symlink_path)
-                    .map_err(|e| Error::Io(symlink_path.clone(), e))?;
-                continue;
-            }
-            // Match the parent's convention: global-store materialization
-            // walks sibling subdirs under their hashed names, while the
-            // per-project `.aube/` layout uses raw dep_paths.
-            let sibling_subdir = if apply_hashes {
-                self.virtual_store_subdir(&dep_dep_path)
-            } else {
-                self.aube_dir_entry_name(&dep_dep_path)
-            };
-            // Compute the relative path from the symlink's parent to
-            // the sibling dep directory. The symlink's parent is
-            // `pkg_nm_parent/` for a bare name but
-            // `pkg_nm_parent/@scope/` for a scoped one, so we can't
-            // hard-code `../..` — doing so would undercount by one
-            // level for every scoped transitive dep and produce a
-            // dangling link. `pathdiff::diff_paths` walks the
-            // difference for us, yielding `../..` for `foo` and
-            // `../../..` for `@vue/shared`, both relative to whatever
-            // parent `symlink_path` ends up with.
-            // `pkg_nm_parent` is `<base_dir>/<subdir>/node_modules/`, so
-            // two parents deep brings us to `<base_dir>/` where all
-            // sibling subdirs live side-by-side.
-            let virtual_root = pkg_nm_parent
-                .parent()
-                .and_then(Path::parent)
-                .unwrap_or(&pkg_nm_parent);
-            let sibling_abs = virtual_root
-                .join(&sibling_subdir)
-                .join("node_modules")
-                .join(dep_name);
-            let link_parent = symlink_path.parent().unwrap_or(&pkg_nm_parent);
-            let target = pathdiff::diff_paths(&sibling_abs, link_parent)
-                .unwrap_or_else(|| sibling_abs.clone());
-
-            // GVS materialize writes into `.tmp-<pid>-<subdir>/`, then
-            // atomic-renames into `self.virtual_store/<subdir>/`. POSIX
-            // symlinks store the relative offset verbatim. Offset stays
-            // invariant under the wrapper rename, so the link resolves
-            // correctly after the move. Windows junctions resolve the
-            // target against `link.parent()` at create time and persist
-            // an absolute path, which binds the junction to the tmp
-            // wrapper. After rename every sibling link dangles into a
-            // gone `.tmp-<pid>-...` path. Fix: on Windows GVS path
-            // (`apply_hashes = true`) rewrite the target to point at
-            // the final virtual store root so the stored absolute path
-            // survives the rename.
-            #[cfg(windows)]
-            let target = if apply_hashes {
-                self.virtual_store
-                    .join(&sibling_subdir)
-                    .join("node_modules")
-                    .join(dep_name)
-            } else {
-                target
-            };
-
-            sys::create_dir_link(&target, &symlink_path)
-                .map_err(|e| Error::Io(symlink_path.clone(), e))?;
-        }
-
-        stats.packages_linked += 1;
-        trace!("materialized {dep_path} ({} files)", index.len());
-        Ok(())
-    }
-
-    /// Hardlink-or-copy a file into a freshly-created destination.
-    /// Assumes `dst` does not exist — callers (`materialize_into`)
-    /// always write into a `.tmp-<pid>-...` staging dir or a
-    /// just-wiped per-project `.aube/<dep_path>`, so the defensive
-    /// `remove_file(dst)` an idempotent variant would need is skipped.
-    /// Eliminates one syscall per linked file (~45k on the medium
-    /// benchmark fixture).
-    pub(crate) fn link_file_fresh(
-        &self,
-        stored: &StoredFile,
-        rel_path: &str,
-        dst: &Path,
-    ) -> Result<(), Error> {
-        #[cfg(target_os = "macos")]
-        const SMALL_FILE_COPY_MAX: u64 = 16 * 1024;
-        let map_io = |e: std::io::Error| classify_link_error(stored, rel_path, dst, e);
-        let missing_source = || Error::MissingStoreFile {
-            store_path: stored.store_path.clone(),
-            rel_path: rel_path.to_string(),
-        };
-        // Track the realized strategy (may differ from `self.strategy` when
-        // a reflink or hardlink falls back to copy) for diagnostic
-        // attribution. Diag emits a `linker.link_<strategy>` event with
-        // the per-file duration so the analyzer can break down link cost
-        // by realized path: reflink (zero-copy CoW), hardlink (zero-cost
-        // metadata link), copy (full byte transfer), or the
-        // small-file-copy short circuit on macOS.
-        let diag_t0 = aube_util::diag::enabled().then(std::time::Instant::now);
-        let realized: &'static str;
-        match self.strategy {
-            LinkStrategy::Reflink => {
-                #[cfg(target_os = "macos")]
-                if matches!(stored.size, Some(size) if size <= SMALL_FILE_COPY_MAX) {
-                    std::fs::copy(&stored.store_path, dst).map_err(map_io)?;
-                    if let Some(t0) = diag_t0 {
-                        aube_util::diag::event(
-                            aube_util::diag::Category::Linker,
-                            "link_macos_small_copy",
-                            t0.elapsed(),
-                            None,
-                        );
-                    }
-                    return Ok(());
-                }
-                if let Err(e) = reflink_copy::reflink(&stored.store_path, dst) {
-                    // Source-missing short-circuit avoids the misleading
-                    // "fell back to copy" trace and the redundant copy
-                    // attempt that would just ENOENT for the same reason.
-                    if !stored.store_path.exists() {
-                        return Err(missing_source());
-                    }
-                    // Fall back to copy on cross-filesystem errors
-                    trace!("reflink failed, falling back to copy: {e}");
-                    std::fs::copy(&stored.store_path, dst).map_err(map_io)?;
-                    realized = "reflink_fallback_copy";
-                } else {
-                    realized = "reflink";
-                }
-            }
-            LinkStrategy::Hardlink => {
-                if let Err(e) = std::fs::hard_link(&stored.store_path, dst) {
-                    if !stored.store_path.exists() {
-                        return Err(missing_source());
-                    }
-                    // Fall back to copy on cross-filesystem errors (EXDEV)
-                    trace!("hardlink failed, falling back to copy: {e}");
-                    std::fs::copy(&stored.store_path, dst).map_err(map_io)?;
-                    realized = "hardlink_fallback_copy";
-                } else {
-                    realized = "hardlink";
-                }
-            }
-            LinkStrategy::Copy => {
-                std::fs::copy(&stored.store_path, dst).map_err(map_io)?;
-                realized = "copy";
-            }
-        }
-
-        if let Some(t0) = diag_t0 {
-            // `realized` is one of seven static strings; matching is
-            // O(1) and the static `&str` keeps the JSONL category compact.
-            let name = match realized {
-                "reflink" => "link_reflink",
-                "reflink_fallback_copy" => "link_reflink_fallback",
-                "hardlink" => "link_hardlink",
-                "hardlink_fallback_copy" => "link_hardlink_fallback",
-                "copy" => "link_copy",
-                "macos_small_copy" => "link_macos_small_copy",
-                _ => "link_unknown",
-            };
-            aube_util::diag::event(aube_util::diag::Category::Linker, name, t0.elapsed(), None);
-        }
-        Ok(())
-    }
-}
-
-/// Translate a copy failure into the most informative linker error.
-/// ENOENT can mean either side of the operation is missing — stat the
-/// source CAS shard to attribute it. A missing shard means the cached
-/// package index is out of sync with the on-disk store, which the
-/// caller can recover from by invalidating the cached index and
-/// re-importing the tarball.
-fn classify_link_error(
-    stored: &StoredFile,
-    rel_path: &str,
-    dst: &Path,
-    err: std::io::Error,
-) -> Error {
-    if err.kind() == std::io::ErrorKind::NotFound && !stored.store_path.exists() {
-        return Error::MissingStoreFile {
-            store_path: stored.store_path.clone(),
-            rel_path: rel_path.to_string(),
-        };
-    }
-    Error::Io(dst.to_path_buf(), err)
-}
-
-/// Best-effort drop the cached package index when materialize discovers
-/// its referenced CAS shard is gone. Callers always surface the original
-/// `MissingStoreFile` error first; this side effect just makes sure the
-/// next install miss `load_index` instead of looping on the same dead
-/// reference. If the cache write fails (e.g. permission error), warn
-/// loudly so the user knows the auto-recovery didn't take and they need
-/// to wipe the index dir by hand (run `aube store path` to find it).
-pub(crate) fn invalidate_stale_index_for_package(store: &aube_store::Store, pkg: &LockedPackage) {
-    match store.invalidate_cached_index(pkg.registry_name(), &pkg.version, pkg.integrity.as_deref())
-    {
-        Ok(true) => debug!("invalidated stale index for {}", pkg.spec_key()),
-        Ok(false) => {}
-        Err(e) => warn!(
-            "failed to invalidate stale index for {}: {e}; manual recovery: rm -rf \"$(aube store path)/index\"",
-            pkg.spec_key()
-        ),
-    }
-}
-
 #[derive(Debug, Default)]
 pub struct LinkStats {
     pub packages_linked: usize,
@@ -2576,182 +236,4 @@ pub struct LinkStats {
     /// `None` means "isolated layout — use the `.aube/<dep_path>`
     /// convention".
     pub hoisted_placements: Option<HoistedPlacements>,
-}
-
-/// Defence in depth for the tarball path-traversal class. The
-/// primary guard lives in `aube_store::import_tarball`, which
-/// refuses malformed entries before they enter the `PackageIndex`.
-/// This helper is the last check before `base.join(key)` is
-/// written through the linker, so an index loaded from a cache
-/// file that predates the store-side validation (or a bug that
-/// lets a traversing key slip past it) still cannot produce a
-/// file outside the package root.
-fn validate_index_key(key: &str) -> Result<(), Error> {
-    if key.is_empty()
-        || key.starts_with('/')
-        || key.starts_with('\\')
-        || key.contains('\0')
-        || key.contains('\\')
-    {
-        return Err(Error::UnsafeIndexKey(key.to_string()));
-    }
-    // Reject any `..` component or Windows drive prefix like `C:`
-    // that would make `Path::join` escape the base.
-    for component in std::path::Path::new(key).components() {
-        match component {
-            std::path::Component::ParentDir
-            | std::path::Component::RootDir
-            | std::path::Component::Prefix(_) => {
-                return Err(Error::UnsafeIndexKey(key.to_string()));
-            }
-            std::path::Component::Normal(os) => {
-                #[cfg(windows)]
-                {
-                    if let Some(s) = os.to_str()
-                        && s.contains(':')
-                    {
-                        return Err(Error::UnsafeIndexKey(key.to_string()));
-                    }
-                }
-                #[cfg(not(windows))]
-                {
-                    let _ = os;
-                }
-            }
-            std::path::Component::CurDir => {}
-        }
-    }
-    Ok(())
-}
-
-/// Decide whether an existing `node_modules/<name>` entry can be left
-/// alone, or must be removed so the caller can recreate it.
-///
-/// Returns `Ok(true)` when a live entry is present and should be
-/// preserved. Returns `Ok(false)` when nothing is there (or a broken
-/// link was reclaimed) and the caller should proceed to create the
-/// entry. `symlink_metadata().is_ok()` on its own treats a dangling
-/// symlink — whose `.aube/<dep_path>/...` target has been deleted — as
-/// "already in place", which silently leaves the project unresolvable.
-///
-/// `sys::create_dir_link` produces a Unix symlink on Unix and an NTFS
-/// junction on Windows. A junction's `file_type().is_symlink()` is
-/// `false`, so we trust the `symlink_metadata().is_ok() && !exists()`
-/// pair to identify "something is at `path` but its target is gone",
-/// and use the same `remove_dir().or_else(remove_file())` fallback
-/// used elsewhere in this file to unlink both shapes.
-/// Reconcile a top-level `node_modules/<name>` entry against the
-/// expected symlink target. Compares the link's *target* — a version
-/// upgrade that leaves `.aube/<old-dep-path>/` resolvable on disk is
-/// correctly classified as stale instead of silently keeping the old
-/// symlink.
-///
-/// - `Ok(true)`  – existing entry is a symlink pointing at
-///   `expected_target`; caller skips creation.
-/// - `Ok(false)` – no entry exists, or a stale entry (wrong target,
-///   dangling symlink, regular directory) has been best-effort
-///   removed; caller should proceed to create the symlink.
-///
-/// Unix and Windows use different comparison strategies because
-/// `create_dir_link` writes the target differently on each platform:
-/// Unix preserves the relative target bytes-for-bytes as a POSIX
-/// symlink, Windows normalizes to an absolute path before calling
-/// `junction::create`. A plain `read_link == expected` check that
-/// works on Unix would miss every warm run on Windows.
-fn reconcile_top_level_link(link_path: &Path, expected_target: &Path) -> Result<bool, Error> {
-    #[cfg(windows)]
-    {
-        // NTFS junctions store normalized absolute targets
-        // (sometimes `\\?\`-prefixed), so comparing against the
-        // relative `pathdiff::diff_paths` output the callers compute
-        // would never match. Compare the canonical forms instead: if
-        // the junction resolves to the same directory
-        // `expected_target` points at, the link is fresh. Anything
-        // else (dangling, wrong target, not a reparse point) falls
-        // through to a best-effort reclaim.
-        //
-        // Canonicalize is ~5 syscalls on NTFS (open reparse, read
-        // reparse data, close, query attrs ×2). With ~1000 top-level
-        // links per warm install that's 5000 syscalls just for
-        // expected_abs. Cache canonical forms keyed by the absolute
-        // path so a second call to the same target returns
-        // immediately.
-        use std::sync::OnceLock;
-        static CANON_CACHE: OnceLock<
-            std::sync::RwLock<std::collections::HashMap<PathBuf, PathBuf>>,
-        > = OnceLock::new();
-        fn cached_canonicalize(p: &Path) -> std::io::Result<PathBuf> {
-            let map = CANON_CACHE.get_or_init(Default::default);
-            if let Some(hit) = map.read().expect("canon cache poisoned").get(p) {
-                return Ok(hit.clone());
-            }
-            let canon = p.canonicalize()?;
-            map.write()
-                .expect("canon cache poisoned")
-                .insert(p.to_path_buf(), canon.clone());
-            Ok(canon)
-        }
-        let expected_abs = if expected_target.is_absolute() {
-            expected_target.to_path_buf()
-        } else {
-            let parent = link_path.parent().unwrap_or_else(|| Path::new(""));
-            parent.join(expected_target)
-        };
-        if let Ok(link_canon) = cached_canonicalize(link_path)
-            && let Ok(exp_canon) = cached_canonicalize(&expected_abs)
-            && link_canon == exp_canon
-        {
-            return Ok(true);
-        }
-        if link_path.symlink_metadata().is_err() {
-            return Ok(false);
-        }
-        match std::fs::remove_dir(link_path).or_else(|_| std::fs::remove_file(link_path)) {
-            Ok(()) => Ok(false),
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(false),
-            Err(e) => Err(Error::Io(link_path.to_path_buf(), e)),
-        }
-    }
-    #[cfg(not(windows))]
-    {
-        match std::fs::read_link(link_path) {
-            Ok(existing) if existing == expected_target => Ok(true),
-            Ok(_) => {
-                // Wrong target — remove the stale symlink so the
-                // caller's `create_dir_link` below doesn't EEXIST.
-                let _ = std::fs::remove_dir(link_path).or_else(|_| std::fs::remove_file(link_path));
-                Ok(false)
-            }
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(false),
-            Err(_) => {
-                // `read_link` failed with EINVAL (entry exists but
-                // isn't a symlink — e.g. a regular directory left by
-                // a prior hoisted install) or another error.
-                // Best-effort reclaim so the create call lands on a
-                // clean slot.
-                let _ =
-                    std::fs::remove_dir_all(link_path).or_else(|_| std::fs::remove_file(link_path));
-                Ok(false)
-            }
-        }
-    }
-}
-
-/// Build a `dep_path → absolute on-disk target` map for every
-/// `LocalSource::Link` in the graph. Returned `None` when the graph
-/// has no link entries (vast majority of installs), so the materialize
-/// hot path can short-circuit without a per-dep lookup.
-pub fn build_nested_link_targets(
-    project_dir: &Path,
-    graph: &LockfileGraph,
-) -> Option<BTreeMap<String, PathBuf>> {
-    let map: BTreeMap<String, PathBuf> = graph
-        .packages
-        .iter()
-        .filter_map(|(dp, pkg)| match pkg.local_source.as_ref() {
-            Some(LocalSource::Link(rel)) => Some((dp.clone(), project_dir.join(rel))),
-            _ => None,
-        })
-        .collect();
-    if map.is_empty() { None } else { Some(map) }
 }
